% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-citytemp.R
\docType{data}
\name{citytemp}
\alias{citytemp}
\title{City temperatures from a year}
\format{A \code{data frame} with 12 observations and 5 variables.}
\usage{
citytemp
}
\description{
This data comes from the \url{http://www.highcharts.com/} examples.
}
\section{Variables}{


\itemize{

 \item \code{month}: The months.
 \item \code{tokyo}: Tokyo's temperatures.
 \item \code{new_york}: New york's temperatures.
 \item \code{berlin}: Berlin's temperatures.
 \item \code{london}: Londo's temperatures.

}
}
\keyword{datasets}

