% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citytemp_long}
\alias{citytemp_long}
\title{City temperatures from a year in long format}
\format{
A \code{data frame} with 36 observations and 3 variables.
}
\usage{
citytemp_long
}
\description{
This data comes from the \url{https://www.highcharts.com/} examples.
}
\section{Variables}{


\itemize{

\item \code{month}: The months.
\item \code{citiy}: City.
\item \code{temp}: Temperatures.
}
}

\keyword{datasets}
