% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{color_classes}
\alias{color_classes}
\title{Function to create \code{dataClasses} argument in \code{hc_colorAxis}}
\usage{
color_classes(breaks = NULL, colors = c("#440154", "#21908C", "#FDE725"))
}
\arguments{
\item{breaks}{A numeric vector}

\item{colors}{A character string of colors (ordered)}
}
\description{
Function to create \code{dataClasses} argument in \code{hc_colorAxis}
}
\examples{

color_classes(c(0, 10, 20, 50))

}
