% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorize}
\alias{colorize}
\title{Create vector of color from vector}
\usage{
colorize(x, colors = c("#440154", "#21908C", "#FDE725"))
}
\arguments{
\item{x}{A numeric, character or factor object.}

\item{colors}{A character string of colors (ordered) to colorize \code{x}}
}
\description{
Create vector of color from vector
}
\examples{

colorize(runif(10))

colorize(LETTERS[rbinom(20, 5, 0.5)], c("#FF0000", "#00FFFF"))

}

