% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorize_vector}
\alias{colorize_vector}
\title{Create vector of color from vector}
\usage{
colorize_vector(x, option = "D")
}
\arguments{
\item{x}{A numeric, character or factor object.}

\item{option}{A character string indicating the colormap option to use.}
}
\description{
Create vector of color from vector
}

