% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-citytemp.R, R/data-favorite_bars.R, R/data-favorite_pies.R
\docType{data}
\name{citytemp}
\alias{citytemp}
\alias{favorite_bars}
\alias{favorite_pies}
\title{City temperatures from a year}
\format{A \code{data frame} with 12 observations and 5 variables.}
\usage{
citytemp

favorite_bars

favorite_pies
}
\description{
This data comes from the \url{http://www.highcharts.com/} examples.

Data from How I met Your Mother: Marshall's Favorite Bars.

Data from How I met Your Mother: Marshall's Favorite Pies
}
\section{Variables}{


\itemize{

 \item \code{month}: The months.
 \item \code{tokyo}: Tokyo's temperatures.
 \item \code{new_york}: New york's temperatures.
 \item \code{berlin}: Berlin's temperatures.
 \item \code{london}: Londo's temperatures.

}



\itemize{

 \item \code{bar}: Bar's name.
 \item \code{percent}: In percentage of awesomeness

}



\itemize{

 \item \code{pie}: Bar's name.
 \item \code{percent}: In percentage of tastiness

}
}
\examples{
data(citytemp)
citytemp
data(favorite_bars)
favorite_bars
data(favorite_pies)
favorite_pies
}

