% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-helpers.R
\name{data_to_sankey}
\alias{data_to_sankey}
\title{Helper to transform data frame for sankey highcharts format}
\usage{
data_to_sankey(data = NULL)
}
\arguments{
\item{data}{A data frame}
}
\description{
Helper to transform data frame for sankey highcharts format
}
\examples{
\dontrun{
library(dplyr)
data(diamonds, package = "ggplot2")

diamonds2 <- select(diamonds, cut, color, clarity)

data_to_sankey(diamonds2)

hchart(data_to_sankey(diamonds2), "sankey", name = "diamonds")
}

}
