% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{date_to_timestamp}
\alias{date_to_timestamp}
\title{Date to Timesstamps}
\usage{
date_to_timestamp(x)
}
\arguments{
\item{x}{Dates vector}
}
\description{
Turn a date vector to \code{timestamp} format
}
\examples{

date_to_timestamp(as.Date(c("2015-05-08", "2015-09-12"), format = "\%Y-\%m-\%d"))

}

