% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df_to_annotations_labels}
\alias{df_to_annotations_labels}
\title{Function to create annotations arguments from a data frame}
\usage{
df_to_annotations_labels(df, xAxis = 0, yAxis = 0)
}
\arguments{
\item{df}{A data frame with \code{x}, \code{y} and \code{text} columns names.}

\item{xAxis}{Index (js 0-based) of the x axis to put the annotations.}

\item{yAxis}{Index (js 0-based) of the y axis to put the annotations.}
}
\description{
Function to create annotations arguments from a data frame
}
\examples{

df <- data.frame(text = c("hi", "bye"), x = c(0, 1), y = c(1, 0))

df_to_annotations_labels(df)
}
