% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-hc.R
\name{export_hc}
\alias{export_hc}
\title{Function to export to jsfiddle}
\usage{
export_hc(hc, filename = NULL)
}
\arguments{
\item{hc}{A \code{A highcarts object}}

\item{filename}{A string}
}
\description{
Function to export to jsfiddle
}

