% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fa_icon}
\alias{fa_icon}
\alias{fa_icon_mark}
\title{Helpers functions to get FontAwesome icons code}
\usage{
fa_icon(iconname = "circle")

fa_icon_mark(iconname = "circle")
}
\arguments{
\item{iconname}{The icon's name}
}
\description{
Helpers functions to get FontAwesome icons code
}
\examples{

fa_icon("car")


fa_icon_mark("car")

fa_icon_mark(iconname = c("car", "plane", "car"))

}

