% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{favorite_bars}
\alias{favorite_bars}
\title{Marshall's Favorite Bars}
\format{A \code{data frame} with 5 observations and 2 variables.}
\usage{
favorite_bars
}
\description{
Data from How I met Your Mother: Marshall's Favorite Bars.
}
\section{Variables}{


\itemize{

\item \code{bar}: Bar's name.
\item \code{percent}: In percentage of awesomeness

}
}

\keyword{datasets}
