% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-favorite_pies.R
\docType{data}
\name{favorite_pies}
\alias{favorite_pies}
\title{Marshall's Favorite Pies}
\format{A \code{data frame} with 5 observations and 2 variables.}
\usage{
favorite_pies
}
\description{
Data from How I met Your Mother: Marshall's Favorite Pies
}
\section{Variables}{


\itemize{

 \item \code{pie}: Bar's name.
 \item \code{percent}: In percentage of tastiness

}
}

\keyword{datasets}
