% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_1_length_data}
\alias{fix_1_length_data}
\title{Function to avoid the jsonlite::auto_unbox default}
\usage{
fix_1_length_data(x)
}
\arguments{
\item{x}{And element, numeric or character}
}
\description{
Function to avoid the jsonlite::auto_unbox default
}
\examples{

hchart("A")

highchart() \%>\%
  hc_add_series(data = 1)
}
