% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highmaps.R
\name{get_data_from_map}
\alias{get_data_from_map}
\title{Auxilar function to get the data inside the map data
The urls are listed in \url{https://code.highcharts.com/mapdata/}.}
\usage{
get_data_from_map(mapdata)
}
\arguments{
\item{mapdata}{A list obtained from \code{\link{download_map_data}}.}
}
\description{
Auxilar function to get the data inside the map data
The urls are listed in \url{https://code.highcharts.com/mapdata/}.
}
\examples{
dta <- download_map_data("https://code.highcharts.com/mapdata/countries/us/us-ca-all.js")
get_data_from_map(dta)
}
\seealso{
\code{\link{download_map_data}}
}

