% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_hc_series_from_df}
\alias{get_hc_series_from_df}
\title{Auxiliar function to get series and options from tidy frame for hchart.data.frame}
\usage{
get_hc_series_from_df(data, type = NULL, ...)
}
\arguments{
\item{data}{A \code{data.frame} object.}

\item{type}{The type of chart. Possible values are line, scatter, point, columnn.}

\item{...}{Aesthetic mappings as \code{x y group color low high}.}
}
\description{
This function is used in hchart.data.frame and hc_add_series_df
}
\examples{

highcharter:::get_hc_series_from_df(iris, type = "point", x = Sepal.Width)
}
