% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-globaltemp.R
\docType{data}
\name{globaltemp}
\alias{globaltemp}
\title{globaltemp}
\format{A \code{data frame} with 1992 observations and 4 variables.}
\source{
\url{http://www.climate-lab-book.ac.uk/2016/spiralling-global-temperatures/}
}
\usage{
globaltemp
}
\description{
Temperature information by years.
}
\section{Variables}{


\itemize{

 \item \code{date}: Date.
 \item \code{lower}: Minimum temperature.
 \item \code{median}: Median temperature.
 \item \code{upper}: Maximum temperature.
}
}
\keyword{datasets}

