% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_annotation}
\alias{hc_add_annotation}
\alias{hc_add_annotations}
\title{Helper to add annotations from data frame or list}
\usage{
hc_add_annotation(hc, ...)

hc_add_annotations(hc, x)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/annotations}.}

\item{x}{A \code{list} or a \code{data.frame} of annotations.}
}
\description{
Helper to add annotations from data frame or list
}
\details{
The \code{x} elements must have \code{xValue} and \code{yValue}
elements
}
