% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{hc_add_dependency_fa}
\alias{hc_add_dependency_fa}
\alias{fa_icon}
\alias{fa_icon_mark}
\title{Helpers functions to get FontAwesome icons code}
\usage{
hc_add_dependency_fa(hc)

fa_icon(iconname = "circle")

fa_icon_mark(iconname = "circle")
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{iconname}{The icon's name}
}
\description{
Helpers functions to get FontAwesome icons code
}
