% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-helpers.R
\name{hc_add_event_point}
\alias{hc_add_event_point}
\alias{hc_add_event_series}
\title{Helpers to use highcharter as input in shiny apps}
\usage{
hc_add_event_point(hc, series = "series", event = "click")

hc_add_event_series(hc, series = "series", event = "click")
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{series}{The name of type of series to apply the event.}

\item{event}{The name of event: click, mouseOut,  mouseOver. See 
\url{http://api.highcharts.com/highcharts/plotOptions.areasplinerange.point.events.select}
for more details.}
}
\description{
When you use highcharter in a shiny app, for example
\code{renderHighcharter("my_chart")} and then use the \code{hc_add_event_point}
function you can get the information of the point when the event is triggered
vía \code{input$my_chart}. So the highcharter can be used as an input.
}

