% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hc_add_serie}
\alias{hc_add_serie}
\title{Adding series to highchart object}
\usage{
hc_add_serie(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#chart}
}

