% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_serie_labels_values}
\alias{hc_add_serie_labels_values}
\title{Shorcut to add series for pie, bar and column charts}
\usage{
hc_add_serie_labels_values(hc, labels, values, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{labels}{A vector of labels.}

\item{values}{A numeric vector. Same length of \code{labels}.}

\item{...}{Other parameters for data series.}
}
\description{
This function delete the actual series in the object and change the \code{chart}
type to \code{scatter}.
}
\examples{

\dontrun{

data(favorite_bars)

hc \%>\% 
  hc_chart(type = "pie") \%>\%
  hc_title(text = "My favorite Bars") \%>\%
  hc_subtitle(text = "(In percentage of awesomeness)") \%>\% 
  hc_tooltip(pointFormat= '{point.percentage:.1f}\%') \%>\% 
  hc_add_serie_labels_values(favorite_bars$bar, favorite_bars$percent,
                             colorByPoint = TRUE) \%>\% 
  hc_legend(enabled = FALSE)

data(favorite_pies)

hc \%>\%
  hc_chart(type = "column") \%>\%  
  hc_title(text = "My favorite Pie") \%>\% 
  hc_subtitle(text = "(In percentage of tastiness)") \%>\% 
  hc_tooltip(pointFormat= '{point.y:.1f}\%') \%>\% 
  hc_add_serie_labels_values(favorite_pies$pie, favorite_pies$percent,
                             colorByPoint = TRUE) \%>\% 
  hc_legend(enabled = FALSE)

}

}

