% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_serie_scatter}
\alias{hc_add_serie_scatter}
\title{Shorcut for create scatter plots}
\usage{
hc_add_serie_scatter(hc, x, y, group = NULL, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector. Same length of \code{x}.}

\item{group}{A vector to split the (x,y) pairs (similiar to fill/color in ggplot).}

\item{...}{Aditional shared arguments for the data series (\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This function helps to create scatter plot from two numerics vectors.
}
\examples{

\dontrun{

highchart() \%>\% 
  hc_add_serie_scatter(cars$speed, cars$dist)
   
highchart() \%>\% 
  hc_add_serie_scatter(mtcars$wt, mtcars$mpg, mtcars$cyl) \%>\% 
  hc_chart(zoomType = "xy") \%>\% 
  hc_title(text = "Motor Trend Car Road Tests") \%>\% 
  hc_xAxis(title = list(text = "Weight"), minorTickInterval = "auto") \%>\% 
  hc_yAxis(title = list(text = "Miles/gallon")) \%>\% 
  hc_tooltip(headerFormat = "<b>{series.name} cylinders</b><br>",
             pointFormat = "{point.x} (lb/1000), {point.y} (miles/gallon)")

}

}

