% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_serie_treemap}
\alias{hc_add_serie_treemap}
\title{Shorcut for create treemaps}
\usage{
hc_add_serie_treemap(hc, tm, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{tm}{A \code{treemap} object}

\item{...}{Aditional shared arguments for the data series (\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This function helps to create hicharts treemaps from \code{treemap} objects
from the package \code{treemap}.
}
\examples{

\dontrun{

library("treemap")

data("business")
tm <- treemap(business, 
             index = c("NACE1", "NACE2"), 
             vSize = "turnover", 
             type = "index")
             
highchart() \%>\% 
 hc_add_serie_treemap(tm, allowDrillToNode = TRUE) \%>\% 
 hc_title(text = "Fictitious Business Statistics Data")

}


}

