% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_series.character}
\alias{hc_add_series.character}
\alias{hc_add_series.factor}
\title{hc_add_series for character and factor objects}
\usage{
\method{hc_add_series}{character}(hc, data, ...)

\method{hc_add_series}{factor}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{character} or \code{factor} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/plotOptions.series}.}
}
\description{
hc_add_series for character and factor objects
}
