% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hc_add_series.data.frame}
\alias{hc_add_series.data.frame}
\title{hc_add_series for data frames objects}
\usage{
\method{hc_add_series}{data.frame}(hc, data, mappings = list(), ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} object.}

\item{mappings}{Mappings, same idea as \code{ggplot2}}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
hc_add_series for data frames objects
}

