% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hc_add_series.geo_list}
\alias{hc_add_series.geo_list}
\title{hc_add_series for geo_list objects}
\usage{
\method{hc_add_series}{geo_list}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{geo_list} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
hc_add_series for geo_list objects
}

