% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_add_series.R
\name{hc_add_series.numeric}
\alias{hc_add_series.numeric}
\title{`hc_add_series` for numeric objects}
\usage{
\method{hc_add_series}{numeric}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A numeric object}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
`hc_add_series` for numeric objects
}

