% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hc_add_series.ts}
\alias{hc_add_series.ts}
\title{hc_add_series for time series objects}
\usage{
\method{hc_add_series}{ts}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A time series \code{ts} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
hc_add_series for time series objects
}

