% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hc_add_series.xts}
\alias{hc_add_series.ohlc}
\alias{hc_add_series.xts}
\title{hc_add_series for xts objects}
\usage{
\method{hc_add_series}{xts}(hc, data, ...)

\method{hc_add_series}{ohlc}(hc, data, type = "candlestick", ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{xts} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}

\item{type}{The way to show the \code{xts} object. Can be 'candlestick' or 'ohlc'.}
}
\description{
hc_add_series for xts objects
}

