% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_df}
\alias{hc_add_series_df}
\title{Shorcut for add data series where data is a dataframe}
\usage{
hc_add_series_df(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A data frame.}

\item{...}{Aditional shared arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
Shorcut for add data series where data is a dataframe
}
\examples{

require("dplyr")
n <- 100
df <- data_frame(
  x = rnorm(n),
  y = x * 2 + rnorm(n),
  z =  x^2,
  color = colorize_vector(x)
  )
  
highchart() \%>\% 
  hc_add_series_df(data = df, type = "bubble")


m <- 100
df2 <- data_frame(
  e = rnorm(m),
  y = e + c(1, diff(e))*0.9,
  x = seq(m),
  low = y - 1*e,
  high = y + 1*e,
  color = colorize_vector(e, "A")
)

highchart() \%>\% 
  hc_add_series_df(data = df2, type = "columnrange")

}

