% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_df_tidy}
\alias{hc_add_series_df_tidy}
\title{Shorcut for tidy data frame using a group variable}
\usage{
hc_add_series_df_tidy(hc, data, group, values, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} object.}

\item{group}{bare column name of categories (one for every series)}

\item{values}{bare column name of values}

\item{...}{Aditional shared arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
Shorcut for tidy data frame using a group variable
}
\examples{

dat <- data.frame(id = c(1,2,3,4,5,6),
                  grp = c("A","A","B","B","C","C"),
                  value = c(10,13,9,15,11,16))

highchart() \%>\% 
 hc_chart(type = "column") \%>\% 
 hc_add_series_df_tidy(data = dat, group = grp, values = value)
  
}

