% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_list}
\alias{hc_add_series_list}
\title{Shorcut for data series from a list of data series}
\usage{
hc_add_series_list(hc, lst)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{lst}{A list of data series.}
}
\description{
Shorcut for data series from a list of data series
}
\examples{

ds <- map(seq(5), function(x){
  list(data = cummean(rnorm(100, 2, 5)), name = x)
})

highchart() \%>\% 
  hc_plotOptions(series = list(marker = list(enabled = FALSE))) \%>\% 
  hc_add_series_list(ds)
  
}

