% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_scatter}
\alias{hc_add_series_scatter}
\title{Shorcut for create scatter plots}
\usage{
hc_add_series_scatter(hc, x, y, z = NULL, color = NULL, label = NULL,
  showInLegend = FALSE, viridis.option = "D", ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector. Same length of \code{x}.}

\item{z}{A numeric vector for size. Same length of \code{x}.}

\item{color}{A vector to color the points.}

\item{label}{A vector to put names in the dots if you enable the datalabels.}

\item{showInLegend}{Logical value to show or not the data in the legend box.}

\item{viridis.option}{Palette to use in case the \code{color}. Argument can not
 be \code{NULL} in the case of use \code{color} argument.
Options are A, B, C, D.}

\item{...}{Aditional shared arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This function helps to create scatter plot from two numerics vectors. Options
argumets like size, color and label for points are added.
}
\examples{

require("dplyr")

hc <- highchart() \%>\% 
  hc_title(text = "Motor Trend Car Road Tests") \%>\% 
  hc_xAxis(title = list(text = "Weight")) \%>\% 
  hc_yAxis(title = list(text = "Miles/gallon"))
   
hc_add_series_scatter(hc, mtcars$wt, mtcars$mpg)

hc_add_series_scatter(hc, mtcars$wt, mtcars$mpg,
                     mtcars$drat)
                     
hc_add_series_scatter(hc, mtcars$wt, mtcars$mpg,
                     mtcars$drat, mtcars$cyl)
                     
hc_add_series_scatter(hc, mtcars$wt, mtcars$mpg,
                     mtcars$drat, mtcars$hp,
                     rownames(mtcars),
                     dataLabels = list(
                       enabled = TRUE,
                       format = "{point.label}"
                       )) \%>\% 
hc_chart(zoomType = "xy") \%>\% 
hc_tooltip(useHTML = TRUE,
           headerFormat = "<table>",
           pointFormat = paste("<tr><th colspan=\\"2\\"><h3>{point.label}</h3></th></tr>",
                               "<tr><th>Weight</th><td>{point.x} lb/1000</td></tr>",
                               "<tr><th>MPG</th><td>{point.y} mpg</td></tr>",
                               "<tr><th>Drat</th><td>{point.z} </td></tr>",
                               "<tr><th>HP</th><td>{point.valuecolor} hp</td></tr>"),
           footerFormat = "</table>")

}

