% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hc_add_theme}
\alias{hc_add_theme}
\title{Add themes to a highchart object}
\usage{
hc_add_theme(hc, hc_thm)
}
\arguments{
\item{hc}{A highchart object}

\item{hc_thm}{A highchart theme object (\code{"hc_theme"} class)}
}
\description{
Add highcharts themes to a highchart object.
}

