% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_annotationsOptions}
\alias{hc_annotationsOptions}
\title{Setting Annotations Options}
\usage{
hc_annotationsOptions(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Options defined in \url{http://www.highcharts.com/plugin-registry/single/17/Annotations}.}
}
\description{
Setting Annotations Options
}
