% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_boost}
\alias{hc_boost}
\title{Setting boost module options to highcharts objects}
\usage{
hc_boost(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Options defined in \url{https://api.highcharts.com/highcharts/boost}.}
}
\description{
Options for the Boost module. The Boost module allows certain series types to
be rendered by WebGL instead of the default SVG. This allows hundreds of
thousands of data points to be rendered in milliseconds. In addition to the
WebGL rendering it saves time by skipping processing and inspection of the
data wherever possible.
}
