% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_colors}
\alias{hc_colors}
\title{colors options for highcharter objects}
\usage{
hc_colors(hc, colors)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{colors}{A vector of colors.}
}
\description{
An array containing the default colors for the chart's series. When
all colors are used, new colors are pulled from the start again.
Default colors can also be set on a series or series.type basis,
see column.colors,
pie.colors.
In styled mode, the colors option doesn't exist. Instead, colors
are defined in CSS and applied either through series or point class
names, or through the chart.colorCount option.
Legacy
In Highcharts 3.x, the default colors were:
colors: \link[='#2f7ed8', '#0d233a', '#8bbc21', '#910000', '#1aadce',
'#492970', '#f28f43', '#77a1e5', '#c42525', '#a6c96a']{'#2f7ed8', '#0d233a', '#8bbc21', '#910000', '#1aadce', '#492970', '#f28f43', '#77a1e5', '#c42525', '#a6c96a'}
}
\details{
In Highcharts 2.x, the default colors were:
colors: \link[='#4572A7', '#AA4643', '#89A54E', '#80699B', '#3D96AE',
'#DB843D', '#92A8CD', '#A47D7C', '#B5CA92']{'#4572A7', '#AA4643', '#89A54E', '#80699B', '#3D96AE', '#DB843D', '#92A8CD', '#A47D7C', '#B5CA92'}

Setting color options to highchart objects

An array containing the default colors for the chart's series. When all
colors are used, new colors are pulled from the start again.
}
\examples{

library(viridisLite)

cols <- viridis(3)
cols <- substr(cols, 0, 7)

highchart() \%>\% 
  hc_add_series(data = sample(1:12)) \%>\% 
  hc_add_series(data = sample(1:12) + 10) \%>\% 
  hc_add_series(data = sample(1:12) + 20) \%>\% 
  hc_colors(cols)

}
