% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hc_credits}
\alias{hc_credits}
\title{Adding credits options to highchart objects}
\usage{
hc_credits(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#credits}.}
}
\description{
\code{highcarter} by default don't put credits label. You can add credits
using these options.
}

