% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-helpers.R
\name{hc_elementId}
\alias{hc_elementId}
\title{Setting \code{elementId}}
\usage{
hc_elementId(hc, id = NULL)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{id}{A string}
}
\description{
Function to modify the \code{id} for the container.
}
\examples{

hchart(rnorm(10)) \%>\%
  hc_elementId("newid")

}
