% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_exporting}
\alias{hc_exporting}
\title{Adding exporting options to highchart object}
\usage{
hc_exporting(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#exporting}
}

