% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hc_get_colors}
\alias{hc_get_colors}
\title{Get default colors for Highcharts theme}
\usage{
hc_get_colors()
}
\description{
Get color used in highcharts charts.
}
\examples{

hc_get_colors()[1:5]

}

