% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_grid.R
\name{hc_grid}
\alias{hc_grid}
\title{Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}}
\usage{
hc_grid(..., ncol = 1, nrow = 1, widths = 1)
}
\arguments{
\item{...}{either individual \code{highchart} objects or a mixture of
    individual \code{highchart} objects
and \code{list}s of \code{highchart} objects.}

\item{ncol}{how many columns in the grid}

\item{nrow}{how many rows in the grid}

\item{widths}{widths of the cells per row}
}
\value{
\code{htmltools} tag with wrapped \code{highchart} objects suitable
        for knitting with \code{echo=FALSE} & \code{results="asis"} or
        rendering in Viewer with \code{html_print}
}
\description{
Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}
}
\note{
\code{hc_grid} does not work in a Shiny context
}
\author{
boB Rudis (hrbrmstr)
}

