% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxies.R
\name{hc_hide}
\alias{hc_hide}
\title{Proxy}
\usage{
hc_hide(proxy, ...)
}
\arguments{
\item{proxy}{An object of class \code{highchartProxy} as returned by \code{\link{highchartProxy}}.}
}
\description{
highcharter proxy example.
}
\examples{
library(shiny)

data("diamonds", package = "ggplot2")

ui <- fluidPage(
  actionButton("polar", "Polar"),
  highchartOutput("hc")
)

server <- function(input, output, session){
  output$hc <- renderHighchart({
    hchart(diamonds$price, color = "#B71C1C", name = "Price") \%>\% 
      hc_title(text = "You can zoom me")
  })
  
  observeEvent(input$polar, {
    highchartProxy("hc") \%>\% 
      hc_hide(
        list(chart = list(polar = TRUE))
      )
  })
}

if(interactive()) shinyApp(ui, server)

}
