% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_labels}
\alias{hc_labels}
\title{Labels options for highcharter objects}
\usage{
hc_labels(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in the Highcharts labels API.}
}
\description{
HTML labels that can be positioned anywhere in the chart area.
This option is deprecated since v7.1.2. Instead, use
annotations that support labels.
}
\examples{

highchart() \%>\%
  hc_add_series(data = sample(1:12)) \%>\%
  hc_labels(
    items = list(
      list(
        html = "<p>Some <b>important</b><br>text</p>",
        style = list(
          left = "150\%",
          top = "150\%"
        )
      )
    )
  )
}
