% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_motion}
\alias{hc_motion}
\title{Adding options to Motion options to highcharts objects}
\usage{
hc_motion(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://github.com/larsac07/Motion-Highcharts-Plugin/wiki}.}
}
\description{
The Motion Highcharts Plugin adds an interactive HTML5 player
to any Highcharts chart (Highcharts, Highmaps and Highstock).
}

