% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_rangeSelector}
\alias{hc_rangeSelector}
\title{Adding scrollbar options to highstock charts}
\usage{
hc_rangeSelector(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highstock#rangeSelector}.}
}
\description{
Options to edit the range selector which is The range selector is a tool
for selecting ranges to display within the chart. It provides buttons
to select preconfigured ranges in the chart, like 1 day, 1 week, 1 month
etc. It also provides input boxes where min and max dates can be manually
input.
}
