% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_rm_serie}
\alias{hc_rm_serie}
\title{Removing series to highchart object}
\usage{
hc_rm_serie(hc, name = NULL)
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#chart}
}

