% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_rm_series}
\alias{hc_rm_serie}
\alias{hc_rm_series}
\title{Removing series to highchart objects}
\usage{
hc_rm_series(hc, name = NULL)

hc_rm_serie(hc, name = NULL)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{name}{The serie's name to delete.}
}
\description{
Removing series to highchart objects
}

