% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_scrollbar}
\alias{hc_scrollbar}
\title{Adding scrollbar options to highstock objects}
\usage{
hc_scrollbar(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highstock#scrollbar}.}
}
\description{
Options regarding the scrollbar which is a means of panning 
over the X axis of a chart.
}

