% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_size}
\alias{hc_size}
\title{Changing the size}
\usage{
hc_size(hc, width = NULL, height = NULL)
}
\arguments{
\item{width}{A numeric input in pixels.}
}
\description{
Changing the size
}
\examples{

hc_size(hcts(rnorm(100)), 400, 200)

}

