% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-538.R, R/theme-alone.R, R/theme-bloom.R,
%   R/theme-chalk.R, R/theme-darkunica.R, R/theme-db.R, R/theme-economist.R,
%   R/theme-elementary.R, R/theme-ffx.R, R/theme-flat.R, R/theme-flatdark.R,
%   R/theme-ft.R, R/theme-ggplot2.R, R/theme-google.R, R/theme-gridlight.R,
%   R/theme-handdrawn.R, R/theme-hcrt.R, R/theme-monokai.R, R/theme-null.R,
%   R/theme-sandsignika.R, R/theme-smpl.R, R/theme-sparkline.R,
%   R/theme-superheroes.R, R/theme-tufte.R
\name{hc_theme_538}
\alias{hc_theme_538}
\alias{hc_theme_alone}
\alias{hc_theme_bloom}
\alias{hc_theme_chalk}
\alias{hc_theme_darkunica}
\alias{hc_theme_db}
\alias{hc_theme_economist}
\alias{hc_theme_elementary}
\alias{hc_theme_ffx}
\alias{hc_theme_flat}
\alias{hc_theme_flatdark}
\alias{hc_theme_ft}
\alias{hc_theme_ggplot2}
\alias{hc_theme_google}
\alias{hc_theme_gridlight}
\alias{hc_theme_handdrawn}
\alias{hc_theme_hcrt}
\alias{hc_theme_monokai}
\alias{hc_theme_null}
\alias{hc_theme_sandsignika}
\alias{hc_theme_smpl}
\alias{hc_theme_sparkline}
\alias{hc_theme_sparkline_vb}
\alias{hc_theme_superheroes}
\alias{hc_theme_tufte}
\alias{hc_theme_tufte2}
\title{Theme collection for highcharts}
\usage{
hc_theme_538(...)

hc_theme_alone(...)

hc_theme_bloom(...)

hc_theme_chalk(...)

hc_theme_darkunica(...)

hc_theme_db(...)

hc_theme_economist(...)

hc_theme_elementary(...)

hc_theme_ffx(...)

hc_theme_flat(...)

hc_theme_flatdark(...)

hc_theme_ft(...)

hc_theme_ggplot2(...)

hc_theme_google(...)

hc_theme_gridlight(...)

hc_theme_handdrawn(...)

hc_theme_hcrt(...)

hc_theme_monokai(...)

hc_theme_null(...)

hc_theme_sandsignika(...)

hc_theme_smpl(...)

hc_theme_sparkline(...)

hc_theme_sparkline_vb(...)

hc_theme_superheroes(...)

hc_theme_tufte(...)

hc_theme_tufte2(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Highcharts is very flexible so you can modify every element of the chart.
There are some exiting themes so you can apply style to charts with few
lines of code.

Chalk theme for highcharts was inspired by
\url{https://www.amcharts.com/inspiration/hand-drawn/}.

Elementary (OS) theme for highcharts was based on \url{https://elementary.io}

Firefox theme was inspired by \url{https://www.mozilla.org/en-US/styleguide/}.

Flat and flatdark theme is inspired by
\url{https://github.com/chriskempson/base16} and
\url{https://github.com/cttobin/ggthemr#flat}

ggplot2 theme is based on \url{https://ggplot2.tidyverse.org/}.

Google theme for highcharts is based on \url{https://books.google.com/ngrams/}.

Hand Drawn theme for highcharts. Inspired by
\url{https://www.amcharts.com/inspiration/hand-drawn/}.

Theme use for the documentatios webiste.

Monokai is a well know text editor theme.

For Null theme the axis are removed (\code{visible = FALSE}).

Theme smpl design is inspired by \url{https://github.com/hrbrmstr/hrbrmisc/blob/master/R/themes.r}
and color by \url{https://www.materialui.co/flatuicolors}.

Sparkline theme is based on \url{http://www.highcharts.com/demo/sparkline}
and this post  \url{http://jkunst.com/blog/posts/2020-06-26-valuebox-and-sparklines/}.

The superheroes theme is inspired by \url{https://public.tableau.com/profile/ryansmith#!/vizhome/HeroesofNewYork/SuperheroesinNewYork}
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_538())


highcharts_demo() \%>\%
  hc_add_theme(hc_theme_alone())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_bloom())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_chalk())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_darkunica())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_db())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_economist())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_elementary())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ffx())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_flat())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_flatdark())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ft())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ggplot2())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_google())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_gridlight())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_handdrawn())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_hcrt())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_monokai())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_null())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_sandsignika())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_smpl())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_sparkline())
  

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_sparkline_vb())


highcharts_demo() \%>\%
  hc_add_theme(hc_theme_superheroes())
  

n <- 15

dta <- data.frame(
  x = 1:n + rnorm(n),
  y = 2 * 1:n + rnorm(n)
)

highchart() \%>\%
  hc_chart(type = "scatter") \%>\%
  hc_add_series(data = list_parse(dta), showInLegend = FALSE) \%>\%
  hc_add_theme(hc_theme_tufte())


highchart() \%>\%
  hc_chart(type = "column") \%>\%
  hc_add_series(data = round(1 + abs(rnorm(12)), 2), showInLegend = FALSE) \%>\%
  hc_xAxis(categories = month.abb) \%>\%
  hc_add_theme(hc_theme_tufte2())

}
