% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-alone.R
\name{hc_theme_alone}
\alias{hc_theme_alone}
\title{Alone theme for highcharts}
\usage{
hc_theme_alone(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Alone theme for highcharts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_alone())
}
