% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-bloom.R
\name{hc_theme_bloom}
\alias{hc_theme_bloom}
\title{Boolmberg Graphics theme for highcharts}
\usage{
hc_theme_bloom(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Boolmberg Graphics theme for highcharts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_bloom())
}
