% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-chalk.R
\name{hc_theme_chalk}
\alias{hc_theme_chalk}
\title{Chalk theme for highcharts}
\usage{
hc_theme_chalk(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Chalk theme for highcharts. Inspirated by https://www.amcharts.com/inspiration/chalk/
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_chalk())

}

