% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-darkunica.R
\name{hc_theme_darkunica}
\alias{hc_theme_darkunica}
\title{Dark Unica theme for highcharts}
\usage{
hc_theme_darkunica(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Dark Unica theme for highcharts
}
\examples{

highcharts_demo() \%>\% 
 hc_add_theme(hc_theme_darkunica())

}

