% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-db.R
\name{hc_theme_db}
\alias{hc_theme_db}
\title{Dotabuff theme for highcharts}
\usage{
hc_theme_db(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Dotabuff theme for highcharts
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_db())

}
