% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-economist.R
\name{hc_theme_economist}
\alias{hc_theme_economist}
\title{Fivethirtyeight theme for highcharts}
\usage{
hc_theme_economist()
}
\description{
Fivethirtyeight theme for highcharts
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_economist())

}

