% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elementary.R
\name{hc_theme_elementary}
\alias{hc_theme_elementary}
\title{Elementary (OS) theme for highcharts}
\usage{
hc_theme_elementary(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Elementary (OS) theme for highcharts based on \url{https://elementary.io}
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_elementary())

}

