% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ffx.R
\name{hc_theme_ffx}
\alias{hc_theme_ffx}
\title{Firefox theme for highcharts}
\usage{
hc_theme_ffx(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Firefox theme was inspired by \url{https://brand.mozilla.com/all-brands}.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ffx())
}
