% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-flatdark.R
\name{hc_theme_flatdark}
\alias{hc_theme_flatdark}
\title{Flatdark theme for highcharts}
\usage{
hc_theme_flatdark(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Base16 inspired theme \url{https://github.com/chriskempson/base16} and
\url{https://github.com/cttobin/ggthemr#flat}
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_flatdark())

}

