% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ft.R
\name{hc_theme_ft}
\alias{hc_theme_ft}
\title{Financial Times theme for highcharts}
\usage{
hc_theme_ft(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Financial Times theme for highcharts
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_ft())

}

