% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-google.R
\name{hc_theme_google}
\alias{hc_theme_google}
\title{Google theme for highcharts}
\usage{
hc_theme_google(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Google theme for highcharts is based on \url{https://books.google.com/ngrams/}.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_google())
}
