% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gridlight.R
\name{hc_theme_gridlight}
\alias{hc_theme_gridlight}
\title{Grid Light theme for highchart}
\usage{
hc_theme_gridlight(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Grid Light theme for highchart
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_gridlight())

}

