% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-hcrt.R
\name{hc_theme_hcrt}
\alias{hc_theme_hcrt}
\title{Highcharter theme for highcharts}
\usage{
hc_theme_hcrt(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
hcrt theme is used for the documentation website.
}
\examples{

highcharts_demo() |>
  hc_add_theme(hc_theme_hcrt())
}
