% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-highcontrast-light.R
\name{hc_theme_highcontrast_light}
\alias{hc_theme_highcontrast_light}
\title{High Contrast Light theme for highcharts
Accessible high-contrast theme for Highcharts tailored
towards 3:1 contrast against white/off-white backgrounds.
Adapted from https://github.com/highcharts/highcharts/blob/master/ts/Extensions/Themes/HighContrastLight.ts}
\usage{
hc_theme_highcontrast_light(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
High Contrast Light theme for highcharts
Accessible high-contrast theme for Highcharts tailored
towards 3:1 contrast against white/off-white backgrounds.
Adapted from https://github.com/highcharts/highcharts/blob/master/ts/Extensions/Themes/HighContrastLight.ts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_highcontrast_light())
}
