% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hc_theme_merge}
\alias{hc_theme_merge}
\title{Merge themes}
\usage{
hc_theme_merge(...)
}
\arguments{
\item{...}{A \code{hc_theme} objects.}
}
\description{
Function to combine hc_theme objects.
}

