% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sandsignika.R
\name{hc_theme_sandsignika}
\alias{hc_theme_sandsignika}
\title{Sand Signika theme for highcharts}
\usage{
hc_theme_sandsignika(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Sand Signika theme for highcharts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_sandsignika())
}
