% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smpl.R
\name{hc_theme_smpl}
\alias{hc_theme_smpl}
\title{Simple theme for highcharts}
\usage{
hc_theme_smpl(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Theme smpl design is inspired by hrbrmisc by hrbrmstr.
and color by \url{https://materialui.co/flatuicolors}.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_smpl())
}
