% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-sparkline.R
\name{hc_theme_sparkline}
\alias{hc_theme_sparkline}
\title{Sparkline theme for highcharts}
\usage{
hc_theme_sparkline(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Based on \url{http://www.highcharts.com/demo/sparkline}.
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_sparkline())
  
}

