% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-superheroes.R
\name{hc_theme_superheroes}
\alias{hc_theme_superheroes}
\title{Superheroes theme for highcharts}
\usage{
hc_theme_superheroes(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
The superheroes theme is inspired by \url{https://public.tableau.com/app/profile/ryansmith#!/vizhome/HeroesofNewYork/SuperheroesinNewYork}
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_superheroes())
}
