% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-tufte.R
\name{hc_theme_tufte}
\alias{hc_theme_tufte}
\title{Tufte theme for highcharts}
\usage{
hc_theme_tufte(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Tufte theme for highcharts
}
\examples{

n <- 15

dta <- data.frame(
  x = 1:n + rnorm(n),
  y = 2 * 1:n + rnorm(n)
)

highchart() |>
  hc_chart(type = "scatter") |>
  hc_add_series(data = list_parse(dta), showInLegend = FALSE) |>
  hc_add_theme(hc_theme_tufte())
}
