% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_time}
\alias{hc_time}
\title{time options for highcharter objects}
\usage{
hc_time(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in https://api.highcharts.com/highcharts/time.}
}
\description{
Time options that can apply globally or to individual charts. These
settings affect how datetime axes are laid out, how tooltips are
formatted, how series
pointIntervalUnit works and how
the Highstock range selector handles time.
The common use case is that all charts in the same Highcharts object
share the same time settings, in which case the global settings are set
using setOptions.
// Apply time settings globally
Highcharts.setOptions({
time: {
timezone: 'Europe/London'
}
});
// Apply time settings by instance
var chart = Highcharts.chart('container', {
time: {
timezone: 'America/New_York'
},
series: [{
data: \link{1, 4, 3, 5}
}]
});
}
\details{
// Use the Time object
console.log(
'Current time in New York',
chart.time.dateFormat('%Y-%m-%d %H:%M:%S', Date.now())
);
Since v6.0.5, the time options were moved from the global obect to the
time object, and time options can be set on each individual chart.
}
\examples{

hc <- highchart() \%>\% 
  hc_add_series(
    data = c(29.9, 71.5, 106.4, 129.2, 144.0, 176.0, 135.6, 148.5, 216.4, 194.1, 95.6, 54.4),
    pointStart = JS("Date.UTC(2017, 0, 1)"),
    pointInterval = 36e5
  ) \%>\% 
  hc_xAxis(
    type = "datetime"
  )

hc \%>\% 
  hc_title(text = "Berlin Time") \%>\% 
  hc_time(timezone = "Europe/Berlin")

hc \%>\% 
  hc_title(text = "New York Time") \%>\% 
  hc_time(timezone = "America/New_York")

}
