% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hc_xAxis}
\alias{hc_xAxis}
\title{Adding xAxis options to highchart object}
\usage{
hc_xAxis(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#xAxis}
}

