% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_yAxis}
\alias{hc_yAxis}
\title{Adding yAxis options to highchart object}
\usage{
hc_yAxis(hc, ...)
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#yAxis}
}

