% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_zAxis}
\alias{hc_zAxis}
\title{zAxis options for highcharter objects}
\usage{
hc_zAxis(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in https://api.highcharts.com/highcharts/zAxis.}
}
\description{
The Z axis or depth axis for 3D plots.
See the Axis class for programmatic
access to the axis.
}
\examples{

highchart() \%>\%
  hc_chart(
    type = "scatter3d",
    options3d = list(
      enabled = TRUE,
      alpha = 20,
      beta = 30,
      depth = 200,
      viewDistance = 5
    )
  ) \%>\% 
  hc_add_series(
    data = list(
      list(
        list(1, 2, 3),
        list(2, 3, 4)
      )
    )
  ) \%>\% 
  hc_zAxis(
    startOnTick = FALSE,
    tickInterval = 2,
    tickLength = 4,
    tickWidth = 1,
    gridLineColor = "#A00000",
    gridLineDashStyle = "dot"  
  )

}
