% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hcaes_string}
\alias{hcaes_string}
\alias{hcaes_}
\title{Define aesthetic mappings using strings.
Similar in spirit to \code{ggplot2::aes_string}}
\usage{
hcaes_string(x, y, ...)

hcaes_(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aesthetics must be named.}
}
\description{
Define aesthetic mappings using strings.
Similar in spirit to \code{ggplot2::aes_string}
}
\examples{
hchart(mtcars, "point", hcaes_string("hp", "mpg", group = "cyl"))

hcaes_string(x = 'xval', color = 'colorvar', group = 'grvar')
}
