% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcboxplot}
\alias{hcboxplot}
\title{Function to make a boxplot}
\usage{
hcboxplot(x, by = NULL, outliers = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{by}{A character vector same length of \code{x}}

\item{outliers}{A boolean value to show or not the outliers}

\item{...}{Aditional shared arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make a boxplot
}

