% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hccoef}
\alias{hccoef}
\title{Function to chart coefficients from glms}
\usage{
hccoef(glm)
}
\arguments{
\item{glm}{A \code{glm} object.}
}
\description{
Function to chart coefficients from glms
}

