% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart.R
\name{hchart.igraph}
\alias{hchart.igraph}
\title{Plot igraph objects using Highcharts}
\usage{
\method{hchart}{igraph}(object, ..., layout = NULL)
}
\arguments{
\item{...}{Additional arguments for the data series
(\url{https://api.highcharts.com/highcharts/series}).}

\item{layout}{A layout from igraph package.}
}
\description{
Plot igraph objects using Highcharts
}
\examples{

if(require("igraph")) {

N <- 40
net <- sample_gnp(N, p = 2 / N)
wc <- cluster_walktrap(net)
V(net)$label <- seq(N)
V(net)$name <- paste("I'm #", seq(N))
V(net)$page_rank <- round(page.rank(net)$vector, 2)
V(net)$betweenness <- round(betweenness(net), 2)
V(net)$degree <- degree(net)
V(net)$size <- V(net)$degree
V(net)$comm <- membership(wc)
V(net)$color <- colorize(membership(wc))
hchart(net, layout = layout_with_fr)

}


}
