% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart.R
\name{hchart.qcc}
\alias{hchart.qcc}
\title{qcc charts.  This is like adding in at least 5 more}
\usage{
\method{hchart}{qcc}(object, ...)
}
\arguments{
\item{object}{the object that is sent in
#@import qcc}
}
\description{
qcc charts.  This is like adding in at least 5 more
}
