% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart.R
\name{hchart}
\alias{hchart}
\title{Function to chart R objects}
\usage{
hchart(x, ...)
}
\arguments{
\item{x}{A R object.}

\item{...}{Aditional arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This is a generic function to plot various object including
character, factors, time series (\code{xts}, \code{ts} 
and \code{mts}).
}

