% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hciconarray}
\alias{hciconarray}
\title{Shortcut to make icon arrays charts}
\usage{
hciconarray(labels, counts, rows = NULL, icons = NULL, size = 4, ...)
}
\arguments{
\item{labels}{A character vector}

\item{counts}{A integer vector}

\item{rows}{A integer to set}

\item{icons}{A character vector same length (o length 1) as labels}

\item{size}{Font size}

\item{...}{Additional arguments for the data series \url{https://api.highcharts.com/highcharts/series}.}
}
\description{
Shortcut to make icon arrays charts
}
