% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcpie}
\alias{hcpie}
\title{Shortcut to make a pie chart}
\usage{
hcpie(x, ...)
}
\arguments{
\item{x}{A character o factor vector.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make a pie chart
}
