% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_add_point}
\alias{hcpxy_add_point}
\title{Add point to a series of a higchartProxy object}
\usage{
hcpxy_add_point(proxy, id = NULL, point, redraw = TRUE, shift = FALSE)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{id}{A character vector indicating the \code{id} of the series to update.}

\item{point}{The point options. If options is a single number, a point with
that y value is appended to the series. If it is an list, it will be
interpreted as x and y values respectively. If it is an object,
advanced options as outlined under series.data are applied}

\item{redraw}{Whether to redraw the chart after the point is added. When
adding more than one point, it is highly recommended that the redraw option
be set to false, and instead Highcharts.Chart#redraw is explicitly called
after the adding of points is finished. Otherwise, the chart will redraw
after adding each point.}

\item{shift}{If \code{TRUE}, a point is shifted off the start of the series as
one is appended to the end.}
}
\description{
Add point to a series of a higchartProxy object
}
