% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_add_series}
\alias{hcpxy_add_series}
\title{Add data to higchartProxy element}
\usage{
hcpxy_add_series(proxy, data = NULL, ...)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{data}{An R object supported by \code{hc_add_series} like data frame, ts, etc.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/plotOptions.series}.}
}
\description{
Add data to higchartProxy element
}
