% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_loading}
\alias{hcpxy_loading}
\title{Show or hide loading text for a higchartProxy object}
\usage{
hcpxy_loading(proxy, action = "show")
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{action}{Single-element character vector indicating to \code{"show"} or
\code{"hide"} the loading text defined in \code{lang} options.}
}
\description{
Show or hide loading text for a higchartProxy object
}
