% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_redraw}
\alias{hcpxy_redraw}
\title{Redraw a higchartProxy object}
\usage{
hcpxy_redraw(proxy)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}
}
\description{
Redraw a higchartProxy object
}
