% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_remove_point}
\alias{hcpxy_remove_point}
\title{Remove point to a series of a higchartProxy object}
\usage{
hcpxy_remove_point(proxy, id = NULL, i = NULL, redraw = TRUE)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{id}{A character vector indicating the \code{id} of the series to update.}

\item{i}{The index of the point in the data array. Remember js is 0 based index.}

\item{redraw}{Whether to redraw the chart after the point is added. When
adding more than one point, it is highly recommended that the redraw option
be set to false, and instead Highcharts.Chart#redraw is explicitly called
after the adding of points is finished. Otherwise, the chart will redraw
after adding each point.}
}
\description{
Remove point to a series of a higchartProxy object
}
