% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_remove_series}
\alias{hcpxy_remove_series}
\title{Remove series to higchartProxy element}
\usage{
hcpxy_remove_series(proxy, id = NULL, all = FALSE)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{id}{A character vector indicating the \code{id} (or \code{id}s) of the series to remove.}

\item{all}{A logical value to indicate to remove or not all series. The values
is used only when the value is \code{TRUE}.}
}
\description{
Remove series to higchartProxy element
}
