% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_update}
\alias{hcpxy_update}
\title{Update options for a higchartProxy object}
\usage{
hcpxy_update(proxy, ...)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{...}{Named options.}
}
\description{
Update options for a higchartProxy object
}
