% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{hcpxy_update_point}
\alias{hcpxy_update_point}
\title{Update options series in a higchartProxy object}
\usage{
hcpxy_update_point(proxy, id = NULL, id_point = NULL, ...)
}
\arguments{
\item{proxy}{A \code{higchartProxy} object.}

\item{id}{A character indicating the \code{id} of the series' point to update.}

\item{id_point}{A vector value indicating the point's index to update, (0 based).}

\item{...}{Arguments defined in \url{https://api.highcharts.com/class-reference/Highcharts.Point}.
The arguments will be the same for each series. So if you want update data it
is used this function sequentially for each point}
}
\description{
Update options series in a higchartProxy object
}
