% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcspark}
\alias{hcspark}
\title{Shorcut to make spkarlines}
\usage{
hcspark(x = NULL, type = NULL, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Type sparkline: line, bar, etc.}

\item{...}{Aditional arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Shorcut to make spkarlines
}
\examples{

set.seed(123)
x <- cumsum(rnorm(10))

hcspark(x) 
hcspark(x, "column")
hcspark(c(1, 4, 5), "pie")
hcspark(x, type = "area")
   
}

