% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcsunburst}
\alias{hcsunburst}
\title{Shortcut to create sunburst chart.}
\usage{
hcsunburst(data, group_vars, size_var, ...)
}
\arguments{
\item{data}{data frame containing variables to organize each level of the treemap on}

\item{group_vars}{vector of strings containing column names of variables to generate treemap levels from. the first listed column will specify the top level of the sunburst chart. the unique values in each of these columns must have no intersection (including NAs).}

\item{size_var}{string name of column containing numeric data to aggregate by}

\item{...}{additional shared arguments for the data series
(\url{http://api.highcharts.com/highcharts#series}).}
}
\value{
highchart plot object
}
\description{
This function helps create highcharts sunburst from data frames.
}
