% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcts}
\alias{hcts}
\title{Shortcut to make time series or line charts}
\usage{
hcts(x, ...)
}
\arguments{
\item{x}{A numeric vector or a time series object.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make time series or line charts
}

