% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-hcworld.R
\docType{data}
\name{hcworld}
\alias{hcworld}
\title{World map in Geojson format (list)}
\format{A \code{list} n geojson format.}
\usage{
hcworld
}
\description{
This data comes from the \url{http://www.highcharts.com/} examples.
}
\examples{
data(hcworld)
hcworld
}

