% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list.parse2}
\alias{hc_get_colors}
\alias{hc_get_dash_styles}
\alias{list.parse2}
\alias{str_to_id}
\title{Convert an object to list with identical structure}
\usage{
list.parse2(df)

str_to_id(x)

hc_get_colors()

hc_get_dash_styles()
}
\arguments{
\item{df}{A data frame to}

\item{x}{A vector string.}
}
\description{
This function is similiar to \code{rlist::list.parse} but this removes names.

Turn a string to \code{id} format used in treemaps.

Get color used in highcharts charts.

Get dash style to use on hichartrs objects.
}

