% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{hex_to_rgba}
\alias{hex_to_rgba}
\title{Transform colors from hexadecimal format to rgba hc notation}
\usage{
hex_to_rgba(x, alpha = 1)
}
\arguments{
\item{x}{colors in hexadecimal format}

\item{alpha}{alpha}
}
\description{
Transform colors from hexadecimal format to rgba hc notation
}
\examples{

hex_to_rgba(x <- c("#440154", "#21908C", "#FDE725"))


}

