% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-package.R
\docType{package}
\name{hicharter}
\alias{hicharter}
\alias{hicharter-package}
\title{An \code{htmlwidget} interface to the
\href{http://www.highcharts.com/}{Highcharts} javascript chart library}
\description{
An \code{htmlwidget} interface to the
\href{http://www.highcharts.com/}{Highcharts} javascript chart library
}
\author{
Joshua Kunst (@jbkunst)
}

