% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{highchart2}
\alias{highchart2}
\alias{highchartzero}
\title{Create a Highcharts chart widget}
\usage{
highchart2(
  hc_opts = list(),
  theme = getOption("highcharter.theme"),
  type = "chart",
  width = NULL,
  height = NULL,
  elementId = NULL,
  google_fonts = getOption("highcharter.google_fonts")
)

highchartzero(
  hc_opts = list(),
  theme = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{hc_opts}{A \code{list} object containing options defined as
\url{http://api.highcharts.com/highcharts}.}

\item{theme}{A \code{hc_theme} class object}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}

\item{google_fonts}{A boolean value. If TRUE (default), adds a reference to the
Google Fonts API to the HTML head, downloading CSS for the font families
defined in the Highcharts theme from https://fonts.googleapis.com. Set to
FALSE if you load your own fonts using CSS.}

\item{debug}{A boolean value if you want to print in the browser console the
parameters given to \code{highchart}.}
}
\description{
This widgets don't support options yet.
}
\details{
This function creates a Highchart chart using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
