% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-ct.R
\name{highchart_ct}
\alias{highchart_ct}
\title{Highcharter Crosstalk Widget}
\usage{
highchart_ct(data = NULL, hc_opts = list(),
  theme = getOption("highcharter.theme"), width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{Data frame}

\item{hc_opts}{A \code{list} object containing options defined as 
\url{http://api.highcharts.com/highcharts}.}

\item{theme}{A \code{hc_theme} class object.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
Highcharter Crosstalk Widget
}

