% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-package.R
\name{highcharter-exports}
\alias{highcharter-exports}
\title{highcharter exported operators}
\description{
The following functions are imported and then re-exported
from the taucharts package to enable use of the magrittr
pipe operator with no additional library calls
}

