% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{highchartOutput}
\alias{highchartOutput}
\title{Widget output function for use in Shiny}
\usage{
highchartOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

