% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{highchartProxy}
\alias{highchartProxy}
\title{Send commands to a Highcharts instance in a Shiny app}
\usage{
highchartProxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{Single-element character vector indicating the output ID of
the chart to modify}

\item{session}{The Shiny session object to which the map belongs; usually
the default value will suffice.}
}
\description{
Send commands to a Highcharts instance in a Shiny app
}
