% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{highchartzero}
\alias{highchartzero}
\title{Create a Highcharts chart widget}
\usage{
highchartzero(
  hc_opts = list(),
  theme = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{hc_opts}{A \code{list} object containing options defined as
\url{https://api.highcharts.com/highcharts/}.}

\item{theme}{A \code{hc_theme} class object.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}
}
\description{
This widgets don't support options yet.
}
\details{
This function creates a Highchart chart using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}
