% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is.hexcolor}
\alias{are.hexcolor}
\alias{is.hexcolor}
\title{Check if a string vector is in hexadecimal color format}
\usage{
is.hexcolor(x)

are.hexcolor(x)
}
\arguments{
\item{x}{A string vectors}
}
\description{
Check if a string vector is in hexadecimal color format
}
\examples{

x <- c("#f0f0f0", "#FFf", "#99990000", "#00FFFFFF")

is.hexcolor(x)

are.hexcolor(x)
}

