% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{is.highchart}
\alias{is.highchart}
\title{Reports whether x is a highchart object}
\usage{
is.highchart(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a highchart object
}

