% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{json_verbatim}
\alias{json_verbatim}
\title{Set the class to "json", so it can be interpreted verbatim by
\code{jsonlite::toJSON}.}
\usage{
json_verbatim(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Set the class to "json", so it can be interpreted verbatim by
\code{jsonlite::toJSON}.
}
\examples{
jsonlite::toJSON(
  x = list(
    click = json_verbatim("function() { alert('click') }")
  ),
  auto_unbox = TRUE,
  json_verbatim = TRUE
)
}
