% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list.parse2}
\alias{list.parse2}
\title{Convert an object to list with identical structure}
\usage{
list.parse2(df)
}
\arguments{
\item{df}{A data frame to}
}
\description{
This function is similiar to \code{rlist::list.parse} but this removes names.
}

