% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_parse}
\alias{list.parse2}
\alias{list.parse3}
\alias{list_parse}
\alias{list_parse2}
\title{Convert an object to list with identical structure}
\usage{
list_parse(df)

list_parse2(df)

list.parse2(df)

list.parse3(df)
}
\arguments{
\item{df}{A data frame to parse to list}
}
\description{
This functions are similiar to \code{rlist::list.parse} but this removes
names.
}
\examples{

x <- data.frame(a=1:3, type=c('A','C','B'), stringsAsFactors = FALSE)

list_parse(x)

list_parse2(x)

}

