% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mountains_panorama}
\alias{mountains_panorama}
\title{Visual comparison of Mountains Panorama}
\format{
A \code{data frame} with 91 observations and 3 variables.
}
\usage{
mountains_panorama
}
\description{
This data comes from the \url{https://www.highcharts.com/} examples:
https://www.highcharts.com/demo/3d-area-multiple
}
\section{Variables}{


\itemize{

\item \code{place}: The place.
\item \code{name}: Name.
\item \code{heigth}: Heigth.

}
}

\keyword{datasets}
