% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{mutate_mapping}
\alias{mutate_mapping}
\title{Modify data frame accoring to mapping}
\usage{
mutate_mapping(data, mapping)
}
\arguments{
\item{data}{A data frame object.}

\item{mapping}{A mapping from \code{hcaes} function.}
}
\description{
Modify data frame accoring to mapping
}
\examples{

mutate_mapping(data = head(mtcars), mapping = hcaes(x = cyl, y = wt + cyl, group = gear))

}

