% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{pokemon}
\format{
A \code{data frame} with 898 observations and 24 variables.
}
\usage{
pokemon
}
\description{
Information about 898 pokemon.
}
\keyword{datasets}
