% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{random_id}
\alias{random_id}
\title{Function to generate iids}
\usage{
random_id(n = 1, length = 10)
}
\arguments{
\item{n}{Number of ids}

\item{length}{Length of ids}
}
\description{
Function to generate iids
}

