% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-stars.R
\docType{data}
\name{stars}
\alias{stars}
\title{stars}
\format{A \code{data frame} with 404 observations and 6 variables.}
\usage{
stars
}
\description{
A sample using by Nadieh Bremer blocks. \url{http://bl.ocks.org/nbremer/eb0d1fd4118b731d069e2ff98dfadc47}.
}
\section{Variables}{


\itemize{

\item \code{bv}: BV
\item \code{absmag}: Magnitude
\item \code{lum}: Luminosity
\item \code{temp}: Temperature
\item \code{radiussun}: Radius
\item \code{distance}: Distance
}
}

\keyword{datasets}
