% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tooltip_table}
\alias{tooltip_table}
\title{Helper for make table in tooltips}
\usage{
tooltip_table(x, y, title = NULL, img = NULL, ...)
}
\arguments{
\item{x}{A string vector with description text}

\item{y}{A string with accesors ex: \code{point.series.name}, 
\code{point.x}}

\item{title}{A title tag with accessor or string}

\item{img}{Img tag}

\item{...}{html attributes for the table element}
}
\description{
Helper to make table in tooltips for the \code{pointFormat}
parameter in \code{hc_tooltip}
}
\examples{

x <- c("Income:", "Genre", "Runtime")
y <- c("$ {point.y}", "{point.series.options.extra.genre}",
       "{point.series.options.extra.runtime}")

tooltip_table(x, y)
hc_add_series_scatter(highchart(), mtcars$wt, mtcars$mpg,
                     mtcars$drat, mtcars$hp, rownames(mtcars)) \%>\% 
hc_chart(zoomType = "xy") \%>\% 
hc_tooltip(useHTML = TRUE,
           headerFormat = "<table>",
           pointFormat = paste("<tr><th colspan=\\"2\\"><h3>{point.label}</h3></th></tr>",
                               "<tr><th>Weight</th><td>{point.x} lb/1000</td></tr>",
                               "<tr><th>MPG</th><td>{point.y} mpg</td></tr>",
                               "<tr><th>Drat</th><td>{point.z} </td></tr>",
                               "<tr><th>HP</th><td>{point.valuecolor} hp</td></tr>"),
           footerFormat = "</table>")

}

