% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemployment}
\alias{unemployment}
\title{US Counties unemployment rate}
\format{
A \code{data.frame} with 3 variables and 3.216 observations.
}
\usage{
unemployment
}
\description{
This data comes from the higcharts and is used in highmaps examples.
}
\section{Variables}{


\itemize{

\item \code{code}: The county code.
\item \code{name}: The county name.
\item \code{value}: The unemployment.

}
}

\keyword{datasets}
