% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgeojson}
\alias{usgeojson}
\title{US States map in Geojson format (list)}
\format{
A \code{list} in geojson format.
}
\usage{
usgeojson
}
\description{
This data comes from the \url{https://code.highcharts.com/mapdata/countries/us/us-all.js}
and is used in highmaps examples.
}
\keyword{datasets}
