% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vaccines.R
\docType{data}
\name{vaccines}
\alias{vaccines}
\title{Vaccines}
\format{A \code{data frame} with 3,876 observations and 3 variables.}
\usage{
vaccines
}
\description{
The number of infected people by Measles, measured over 70-some years and across 
all 50 states. From the WSJ analysis: \url{http://graphics.wsj.com/infectious-diseases-and-vaccines/}
}
\section{Variables}{


\itemize{

 \item \code{year}: Year
 \item \code{state}: Name of the state
 \item \code{count}: Number of cases per 100,000 people. If the value is NA the count was 0.

}
}

\keyword{datasets}
