% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-weather.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather}
\format{A \code{data frame} with 365 observations and 4 variables.}
\usage{
weather
}
\description{
Temperature information of San Francisco.
}
\section{Variables}{


\itemize{

 \item \code{date}: Day in date format.
 \item \code{min_temperaturec}: Minimum temperature.
 \item \code{max_temperaturec}: Maximun temperature.
 \item \code{mean_temperaturec}: Mean temperature.
}
}
\keyword{datasets}

