% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{worldgeojson}
\alias{worldgeojson}
\title{World map in Geojson format (list)}
\format{A \code{list} in geojson format.}
\usage{
worldgeojson
}
\description{
This data comes from the \url{https://code.highcharts.com/mapdata/custom/world.js}
and is used in highmaps examples.#'
}
\keyword{datasets}
