% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{color_stops}
\alias{color_stops}
\title{Function to create \code{stops} argumente in \code{hc_colorAxis}}
\usage{
color_stops(n = 10, option = "D")
}
\arguments{
\item{n}{A numeric indicating how much quantiles generate.}

\item{option}{A character string indicating the colormap option to use.}
}
\description{
Using viridis
}
\examples{

color_stops(5)

}

