% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{colorize_vector}
\alias{colorize_vector}
\title{Create vector of color from vector}
\usage{
colorize_vector(x, option = "D")
}
\arguments{
\item{x}{A numeric, character or factor vector.}

\item{option}{A character string indicating the colormap option to use.}
}
\description{
Using viridis
}
\examples{
x <- runif(50)
colorize_vector(x)

x <- sample(letters[1:4], size = 20, replace = TRUE)
colorize_vector(x)

}

