% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_hc.R
\name{export_hc}
\alias{export_hc}
\title{Function to export js file the configuration options}
\usage{
export_hc(hc, filename = NULL)
}
\arguments{
\item{hc}{A \code{A highcarts object}}

\item{filename}{A string}
}
\description{
Function to export js file the configuration options
}
\examples{

fn <- "function(){
  console.log('Category: ' + this.category);
  alert('Category: ' + this.category);
}"

hc <- highcharts_demo() \%>\% 
  hc_plotOptions(
    series = list(
      cursor = "pointer",
        point = list(
          events = list(
            click = JS(fn)
         )
       )
   )
 )

}

