% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_accessibility}
\alias{hc_accessibility}
\title{Setting accessibility options to highcharts objects}
\usage{
hc_accessibility(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Options defined in \url{http://api.highcharts.com/highcharts/accessibility}.}
}
\description{
Options for configuring accessibility for the chart. Requires the accessibility module to be loaded.
}
