% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_series.data.frame}
\alias{hc_add_series.data.frame}
\title{hc_add_series for data frames objects}
\usage{
\method{hc_add_series}{data.frame}(hc, data, type = NULL, mapping = hcaes(), fast = FALSE, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} object.}

\item{type}{The type of the series: line, bar, etc.}

\item{mapping}{The mapping, same idea as \code{ggplot2}.}

\item{fast}{convert to json during the composition of a highchart object}

\item{...}{Arguments defined in
\url{https://api.highcharts.com/highcharts/chart}.}
}
\description{
hc_add_series for data frames objects
}
