% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_add_series.R
\name{hc_add_series.forecast}
\alias{hc_add_series.forecast}
\title{hc_add_series for forecast objects}
\usage{
\method{hc_add_series}{forecast}(hc, data, addOriginal = FALSE,
  addLevels = TRUE, fillOpacity = 0.1, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{forecast} object.}

\item{addOriginal}{Logical value to add the original series or not.}

\item{addLevels}{Logical value to show predicctions bands.}

\item{fillOpacity}{The opacity of bands}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
hc_add_series for forecast objects
}

