% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_series.forecast}
\alias{hc_add_series.forecast}
\title{hc_add_series for forecast objects}
\usage{
\method{hc_add_series}{forecast}(
  hc,
  data,
  addOriginal = FALSE,
  addLevels = TRUE,
  fillOpacity = 0.1,
  name = NULL,
  ...
)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{forecast} object.}

\item{addOriginal}{Logical value to add the original series or not.}

\item{addLevels}{Logical value to show predictions bands.}

\item{fillOpacity}{The opacity of bands.}

\item{name}{The name of the series.}

\item{...}{Arguments defined in
\url{http://api.highcharts.com/highcharts#chart}.}
}
\description{
hc_add_series for forecast objects
}
