% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series.R
\name{hc_add_series.geo_json}
\alias{hc_add_series.geo_json}
\title{hc_add_series for geo_json objecst}
\usage{
\method{hc_add_series}{geo_json}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{geo_json} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts#chart}.}

\item{type}{The type of the series: line, bar, etc.}

\item{mapping}{The mapping, same idea as \code{ggplot2}.}
}
\description{
hc_add_series for geo_json objecst
}

