% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_series.numeric}
\alias{hc_add_series.numeric}
\title{\code{hc_add_series} for numeric objects}
\usage{
\method{hc_add_series}{numeric}(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A numeric object}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/plotOptions.series}.}
}
\description{
\code{hc_add_series} for numeric objects
}
