% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_add_series.xts}
\alias{hc_add_series.xts}
\alias{hc_add_series.ohlc}
\title{hc_add_series for xts objects}
\usage{
\method{hc_add_series}{xts}(hc, data, ...)

\method{hc_add_series}{ohlc}(hc, data, type = "candlestick", ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{xts} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/plotOptions.series}.}

\item{type}{The way to show the \code{xts} object. Can be 'candlestick' or 'ohlc'.}
}
\description{
hc_add_series for xts objects
}
\examples{

if(require("quantmod")){
  x <- getSymbols("GOOG", auto.assign = FALSE)
  hchart(x) 
}

}
