% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_df}
\alias{hc_add_series_df}
\title{Shorcut for add data series where data is a dataframe}
\usage{
hc_add_series_df(hc, data, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A data frame usually with columns: \code{x}, \code{y},
\code{z}, \code{color}, \code{name}, \code{low}, \code{high}.
\url{http://api.highcharts.com/highcharts#series<bubble>.data}}

\item{...}{Aditional shared arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
Automatically parsed de data frame (to a list) so you can use the 
default parameters of highcharts such as \code{x}, \code{y},
\code{z}, \code{color}, \code{name}, \code{low}, \code{high} for 
each series \url{http://api.highcharts.com/highcharts#series<bubble>.data}.
}
\examples{

require("dplyr")
n <- 100
df <- data_frame(
  x = rnorm(n),
  y = x * 2 + rnorm(n),
  z =  x^2,
  color = colorize_vector(x)
  )
  
highchart() \%>\% 
  hc_add_series_df(data = df, type = "bubble")


m <- 100
s <- cumsum(rnorm(m))
e <- 2 + rbeta(m, 2, 2)

df2 <- data_frame(
  x = seq(m),
  low = s - e,
  high = s + e,
  name = paste("I'm point #\%s", x),
  color = colorize_vector(high, "B")
)

highchart() \%>\%
  hc_tooltip(valueDecimals = 2) \%>\% 
  hc_add_series_df(data = df2, name = "I'm a columnrage series",
                   type = "columnrange", showInLegend = FALSE)

}

