% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_df_tidy}
\alias{hc_add_series_df_tidy}
\alias{hcdy}
\title{Shorcut for tidy data frame a la ggplot2/qplot}
\usage{
hc_add_series_df_tidy(hc, data, type = NULL, ...)

hcdy(data, type = NULL, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{data}{A \code{data.frame} object.}

\item{type}{The type of chart}

\item{...}{Aesthetic mappings, \code{x y group color low high}.}
}
\description{
Function to create chart from tidy data frames.
}
\details{
The types supported are line, column, point, polygon,
columrange, spline, areaspline among others
}
\examples{

hcdy(iris, "point", x = Sepal.Length, y = Sepal.Width, group = Species)

dat <- data.frame(id = c(1,2,3,4,5,6),
                  grp = c("A","A","B","B","C","C"),
                  value = c(10,13,9,15,11,16))

highchart() \%>\% 
  hc_add_series_df_tidy(data = dat, type = "column", y = value, group = grp)
  
data(mpg, package = "ggplot2")

hcdy(mpg, "scatter", x = displ, y = cty)
hcdy(mpg, "scatter", x = displ, y = cty, group = manufacturer)
hcdy(mpg, "scatter", x = displ, y = cty, size = hwy, group = manufacturer)
hcdy(mpg, "scatter", x = displ, y = cty, size = hwy, color = class,
     group = manufacturer)
     
require(dplyr)

mpgman <- count(mpg, manufacturer)
hcdy(mpgman, "column", x = manufacturer, y = n)

mpgman2 <- count(mpg, manufacturer, year)
hcdy(mpgman2, "bar", x = manufacturer, y = n, group = year)

data(economics, package = "ggplot2")

hcdy(economics, "line", x = date, y = unemploy)

data(economics_long, package = "ggplot2")

economics_long2 <- filter(economics_long,
                          variable \%in\% c("pop", "uempmed", "unemploy"))
hcdy(economics_long2, "line", x = date, y = value01, group = variable)


}

