% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{hc_add_series_map}
\alias{hc_add_serie_map}
\alias{hc_add_series_map}
\title{Shorcut for create map}
\usage{
hc_add_series_map(hc, map, df, value, joinBy, ...)

hc_add_serie_map(hc, map, df, value, joinBy, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{map}{A \code{list} object loaded from a geojson file.}

\item{df}{A \code{data.frame} object with data to chart. Code region and value are
required.}

\item{value}{A string value with the name of the column to chart.}

\item{joinBy}{What property to join the  \code{map} and \code{df}}

\item{...}{Aditional shared arguments for the data series
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This function helps to create higcharts treemaps from \code{treemap} objects
from the package \code{treemap}.
}

