% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-hc_add_series-shortcuts.R
\name{hc_add_series_map}
\alias{hc_add_series_map}
\title{Shorcut for create map}
\usage{
hc_add_series_map(hc, map, df, value, joinBy, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{map}{A \code{list} object loaded from a geojson file.}

\item{df}{A \code{data.frame} object with data to chart. Code region and value are
required.}

\item{value}{A string value with the name of the column to chart.}

\item{joinBy}{What property to join the  \code{map} and \code{df}}

\item{...}{Aditional shared arguments for the data series
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
This function helps to create higcharts treemaps from \code{treemap} objects
from the package \code{treemap}.
}
\details{
This function force the highchart object to be map type.
}
\examples{

library("dplyr")
library("viridisLite")

data("USArrests", package = "datasets")
data("usgeojson")

USArrests <- USArrests \%>\% 
  mutate(state = rownames(.))

n <- 4
colstops <- data.frame(q = 0:n/n,
                       c = substring(viridis(n + 1, option = "A"), 0, 7)) \%>\% 
list_parse2()

highchart() \%>\% 
  hc_title(text = "Violent Crime Rates by US State") \%>\% 
  hc_subtitle(text = "Source: USArrests data") \%>\% 
  hc_add_series_map(usgeojson, USArrests, name = "Murder arrests (per 100,000)",
                    value = "Murder", joinBy = c("woename", "state"),
                    dataLabels = list(enabled = TRUE,
                                      format = '{point.properties.postalcode}')) \%>\% 
  hc_colorAxis(stops = colstops) \%>\% 
  hc_legend(valueDecimals = 0, valueSuffix = "\%") \%>\%
  hc_mapNavigation(enabled = TRUE) 

\dontrun{

data(worldgeojson, package = "highcharter")
data("GNI2014", package = "treemap")

highchart(type = "map") \%>\% 
  hc_add_series_map(map = worldgeojson, df = GNI2014, value = "GNI", joinBy = "iso3") \%>\% 
  hc_colorAxis(stops = color_stops()) \%>\% 
  hc_tooltip(useHTML = TRUE, headerFormat = "",
  pointFormat = "this is {point.name} and have {point.population} people with gni of {point.GNI}")
  
 }
  
}

