% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hc_add_theme}
\alias{hc_add_theme}
\title{Add themes to a highchart object}
\usage{
hc_add_theme(hc, hc_thm)
}
\arguments{
\item{hc}{A highchart object}

\item{hc_thm}{A highchart theme object (\code{"hc_theme"} class)}
}
\description{
Add highcharts themes to a highchart object.
}
\examples{

highchart() \%>\% 
  hc_add_series(data = c(7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2,
                        26.5, 23.3, 18.3, 13.9, 9.6),
                        type = "column") \%>\% 
  hc_add_theme(hc_theme_sandsignika())
  
}

