% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-helpers.R
\name{hc_add_yAxis}
\alias{hc_add_yAxis}
\title{yAxis add highcharter objects}
\usage{
hc_add_yAxis(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/yAxis}.}
}
\description{
The Y axis or value axis. Normally this is the vertical axis,
though if the chart is inverted this is the horizontal axis.
Add yAxis allows to add multiple axis with a relative height between Y axis.
Based upon the \code{relative} parameter the height of each Y axis is recalculated.
Otherwise the parameters are as supported by Y axis.
}
\examples{

# Retrieve stock data to plot.
aapl <- quantmod::getSymbols("AAPL", 
  src = "yahoo",
  from = "2020-01-01",
  auto.assign = FALSE
)

# Plot prices and volume with relative height.
hc <- highcharter::highchart(type = "stock") \%>\%
  highcharter::hc_title(text = "AAPLE") \%>\%
  highcharter::hc_add_series(aapl, yAxis = 0, showInLegend = FALSE) \%>\%
  highcharter::hc_add_yAxis(nid = 1L, title = list(text = "Prices"), relative = 2) \%>\%
  highcharter::hc_add_series(aapl[, "AAPL.Volume"], yAxis = 1, type = "column", showInLegend = FALSE) \%>\%
  highcharter::hc_add_yAxis(nid = 2L, title = list(text = "Volume"), relative = 1)
  
}
