% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_annotations}
\alias{hc_annotations}
\alias{hc_annotationsOptions}
\title{Adding annotations to highcharts objects}
\usage{
hc_annotations(hc, ...)

hc_annotationsOptions(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://www.highcharts.com/plugin-registry/single/17/Annotations}.}
}
\description{
Helper function to add annotations to highcharts library.
}

