% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_annotations}
\alias{hc_annotations}
\alias{hc_add_annotation}
\alias{hc_add_annotations}
\title{Setting annotations to highcharts objects}
\usage{
hc_annotations(hc, ...)

hc_add_annotation(hc, ...)

hc_add_annotations(hc, x)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/annotations}.}

\item{x}{A \code{list} or a \code{data.frame} of annotations.}
}
\description{
Helper function to add annotations to highcharts library.
}
\details{
The \code{x} elements must have \code{xValue} and \code{yValue}
elements
}
