% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_boost}
\alias{hc_boost}
\title{boost options for highcharter objects}
\usage{
hc_boost(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in https://api.highcharts.com/highcharts/boost.}
}
\description{
Options for the Boost module. The Boost module allows certain series types
to be rendered by WebGL instead of the default SVG. This allows hundreds of
thousands of data points to be rendered in milliseconds. In addition to the
WebGL rendering it saves time by skipping processing and inspection of the
data wherever possible. This introduces some limitations to what features are
available in Boost mode. See the docs for
details.
In addition to the global boost option, each series has a
boostThreshold that defines when the
boost should kick in.
Requires the modules/boost.js module.
}
\examples{

#  Ex 1
n <- 50000

x <- sin(4*2*pi*seq(n)/n) + rnorm(n)/10

x <- round(x, 3)

plot(x)

highchart() \%>\%
  hc_chart(zoomType = "x") \%>\%
  hc_add_series(data = x) \%>\% 
  hc_title(text = "No boost") \%>\% 
  hc_boost(
    enabled = FALSE # Not recommended
  )

highchart() \%>\%
  hc_chart(zoomType = "x") \%>\%
  hc_add_series(data = x) \%>\% 
  hc_title(text = "With boost") \%>\% 
  hc_boost(
    enabled = TRUE # Default :D
  )

# Ex 2
library(MASS)

n <- 100000

sigma <- matrix(c(10,3,3,2),2,2)
sigma

mvr <- round(mvrnorm(n, rep(c(0, 0)), sigma), 6)

vx <- ceiling(1+abs(max(mvr[, 1])))
vy <- ceiling(1+abs(max(mvr[, 2])))

ds <- list_parse2(as.data.frame(mvr))

highchart() \%>\%
  hc_chart(zoomType = "xy") \%>\%
  hc_xAxis(min = -vx, max = vx) \%>\% 
  hc_yAxis(min = -vy, max = vy) \%>\% 
  hc_add_series(
    data = ds,
    type = "scatter",
    name = "A lot of points!",
    color = 'rgba(0,0,0,0.01)',
    marker = list(radius = 2),
    tooltip = list(
      followPointer = FALSE,
      pointFormat =  '[{point.x:.1f}, {point.y:.1f}]'
      )
    )

}
