% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart-api.R
\name{hc_chart}
\alias{hc_chart}
\title{Adding chart options to highchart object}
\usage{
hc_chart(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#chart}
}

