% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hc_colorAxis}
\alias{hc_colorAxis}
\title{Adding Color Axis options to highchart objects}
\usage{
hc_colorAxis(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments are defined in \url{http://www.highcharts.com/docs/maps/color-axis}.}
}
\description{
Adding Color Axis options to highchart objects
}

