% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_colors}
\alias{hc_colors}
\title{Adding color options to highchart objects}
\usage{
hc_colors(hc, colors)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{colors}{A vector of colors.}
}
\description{
An array containing the default colors for the chart's series. When all
colors are used, new colors are pulled from the start again.
}
\examples{

library("viridisLite")

cols <- viridis(3)
cols <- substr(cols, 0, 7)

highcharts_demo() \%>\% 
  hc_colors(cols)


}
