% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_credits}
\alias{hc_credits}
\title{Adding credits options to highchart object}
\usage{
hc_credits(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defided in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#credits}
}

