% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_credits}
\alias{hc_credits}
\title{Credits options for highcharter objects}
\usage{
hc_credits(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/credits}.}
}
\description{
Highchart by default puts a credits label in the lower right corner
of the chart. This can be changed using these options.
}
\examples{

highchart() \%>\% 
  hc_xAxis(categories = citytemp$month) \%>\% 
  hc_add_series(name = "Tokyo", data = sample(1:12)) \%>\% 
  hc_credits(
    enabled = TRUE,
    text = "htmlwidgets.org",
    href = "http://www.htmlwidgets.org/"
    )

}
