% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_defs}
\alias{hc_defs}
\title{Adding patterns to be used in highcharts series}
\usage{
hc_defs(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://www.highcharts.com/plugin-registry/single/9/Pattern-Fill}.}
}
\description{
Helper function to use the fill patter plugin \url{http://www.highcharts.com/plugin-registry/single/9/Pattern-Fill}.
}

