% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_exporting}
\alias{hc_exporting}
\title{Exporting options for highcharter objects}
\usage{
hc_exporting(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highcharts/exporting}.}
}
\description{
Options for the exporting module. For an overview on the matter, see
the docs and
read our Fair Usage Policy.
}
\examples{

highchart() |> 
  hc_xAxis(categories = month.abb) |> 
  hc_add_series(name = "Tokyo", data = sample(1:12)) |> 
  hc_exporting(
    enabled = TRUE, # always enabled
    filename = "custom-file-name"
    )

}
