% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_labels}
\alias{hc_labels}
\title{Adding HTML labels to highcharts objects}
\usage{
hc_labels(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Options defined in \url{https://api.highcharts.com/highcharts/labels}.}
}
\description{
Adds a HTML labels to anywhere in the chart area.
}
\examples{

data(citytemp)

hc <- highchart() \%>\%
  hc_xAxis(categories = citytemp$month) \%>\%
  hc_add_series(name = "Tokyo", data = citytemp$tokyo) \%>\%
  hc_add_series(name = "London", data = citytemp$london) \%>\%
  hc_labels(items=list(
     list(html="This is a label",
        style=list(left="20px",top="20px")
     )
  ))

}
