% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_mapNavigation}
\alias{hc_mapNavigation}
\title{Adding mapNavigation options to highstock charts}
\usage{
hc_mapNavigation(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highmaps#mapNavigation}.}
}
\description{
Options regarding the mapNavigation: A collection of options for zooming 
and panning in a map.
}

