% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_mapNavigation}
\alias{hc_mapNavigation}
\title{Mapnavigation options for highcharter objects}
\usage{
hc_mapNavigation(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{https://api.highcharts.com/highmaps/mapNavigation}.}
}
\description{
Mapnavigation options for highcharter objects
}
\examples{

hcmap(download_map_data = FALSE) \%>\%
  hc_mapNavigation(
    enabled = TRUE,
    enableMouseWheelZoom = TRUE,
    enableDoubleClickZoom = TRUE
  )
}
