% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_motion}
\alias{hc_motion}
\title{Setting Motion options to highcharts objects}
\usage{
hc_motion(hc, enabled = TRUE, startIndex = 0, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{enabled}{Enable the motion plugin.}

\item{startIndex}{start index, default to 0.}

\item{...}{Arguments defined in \url{https://github.com/larsac07/Motion-Highcharts-Plugin/wiki}.}
}
\description{
The Motion Highcharts Plugin adds an interactive HTML5 player
to any Highcharts chart (Highcharts, Highmaps and Highstock).
}
