% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_navigator}
\alias{hc_navigator}
\title{Adding navigator options to highstock charts}
\usage{
hc_navigator(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highstock#navigator}.}
}
\description{
Options regarding the navigator: The miniseries below chart 
in a highstock chart.
}

