% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc.R
\name{hc_pane}
\alias{hc_pane}
\title{Setting panes options to highchart objects}
\usage{
hc_pane(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts/pane}.}
}
\description{
Applies only to polar charts and angular gauges. This configuration object
holds general options for the combined X and Y axes set. Each xAxis or
yAxis can reference the pane by index.
}
