% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api.R
\name{hc_pane}
\alias{hc_pane}
\title{pane options for highcharter objects}
\usage{
hc_pane(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in https://api.highcharts.com/highcharts/pane.}
}
\description{
The pane serves as a container for axes and backgrounds for circular
gauges and polar charts.
}
\examples{



}
