% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_plotOptions}
\alias{hc_plotOptions}
\title{Adding plot options to highchart object}
\usage{
hc_plotOptions(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#plotOptions}
}

