% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_plotOptions}
\alias{hc_plotOptions}
\title{Adding plot options to highchart object}
\usage{
hc_plotOptions(hc, ...)
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#plotOptions}
}

