% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-add.R
\name{hc_rm_series}
\alias{hc_rm_series}
\title{Removing series to highchart objects}
\usage{
hc_rm_series(hc, names = NULL)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{names}{The series's names to delete.}
}
\description{
Removing series to highchart objects
}
