% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-hc-extras.R
\name{hc_size}
\alias{hc_size}
\title{Changing the size of a \code{highchart} object}
\usage{
hc_size(hc, width = NULL, height = NULL)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}
}
\description{
Changing the size of a \code{highchart} object
}
\examples{

hc_size(hcts(rnorm(100)), 400, 200)

}

