% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{hc_subtitle}
\alias{hc_subtitle}
\title{Adding subtitles options to highchart object}
\usage{
hc_subtitle(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defided in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#subtitle}
}

