% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-538.R
\name{hc_theme_538}
\alias{hc_theme_538}
\title{Fivethirtyeight theme for highcharts}
\usage{
hc_theme_538()
}
\description{
Fivethirtyeight theme for highcharts
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_538())

}

