% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-538.R
\name{hc_theme_538}
\alias{hc_theme_538}
\title{Fivethirtyeight theme for highcharts}
\usage{
hc_theme_538()
}
\description{
Fivethirtyeight theme for highcharts
}
\examples{

highchart() \%>\% 
  hc_add_series(data = list.parse2(data.frame(rnorm(15), runif(15))), type = "scatter") \%>\% 
  hc_add_series(data = list.parse2(data.frame(rexp(15), rexp(15))), type = "scatter") \%>\% 
  hc_add_theme(hc_theme_538()) \%>\% 
  hc_title(text = "Some Long Title") \%>\% 
  hc_subtitle(text = "Some subtitle")

}

