% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-538.R, R/theme-sparkline.R,
%   R/theme-tufte.R
\name{hc_theme_538}
\alias{hc_theme_538}
\alias{hc_theme_sparkline_vb}
\alias{hc_theme_tufte2}
\title{Theme collection for highcharts}
\usage{
hc_theme_538(...)

hc_theme_sparkline_vb(...)

hc_theme_tufte2(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Highcharts is very flexible so you can modify every element of the chart.
There are some exiting themes so you can apply style to charts with few
lines of code.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_538())

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_sparkline_vb())

highchart() \%>\%
  hc_chart(type = "column") \%>\%
  hc_add_series(data = round(1 + abs(rnorm(12)), 2), showInLegend = FALSE) \%>\%
  hc_xAxis(categories = month.abb) \%>\%
  hc_add_theme(hc_theme_tufte2())
}
