% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-538.R
\name{hc_theme_538}
\alias{hc_theme_538}
\title{Fivethirtyeight theme for highcharts}
\usage{
hc_theme_538(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Fivethirtyeight theme for highcharts
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_538())
   
highcharts_demo() \%>\%
  hc_add_theme(
    hc_theme_538(
      colors = c("red", "blue", "green"),
      chart = list(backgroundColor = "white")
    )
  )

}
