% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-bloom.R
\name{hc_theme_bloom}
\alias{hc_theme_bloom}
\title{Bloomberg Graphics theme for highcharts}
\usage{
hc_theme_bloom(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Bloomberg Graphics theme for highcharts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_bloom())
  
}
