% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-chalk.R
\name{hc_theme_chalk}
\alias{hc_theme_chalk}
\title{Chalk theme for highcharts}
\usage{
hc_theme_chalk(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.

Chalk theme for highcharts was inspired by
\url{https://www.amcharts.com/demos/}.}
}
\description{
Chalk theme for highcharts
}
\examples{

highcharts_demo() |>
  hc_add_theme(hc_theme_chalk())
}
