% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-economist.R
\name{hc_theme_economist}
\alias{hc_theme_economist}
\title{Economist theme for highcharts}
\usage{
hc_theme_economist(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Economist theme for highcharts
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_economist())
}
