% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ffx.R
\name{hc_theme_ffx}
\alias{hc_theme_ffx}
\title{Firefox}
\usage{
hc_theme_ffx(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Firefox
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_ffx())

}

