% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ffx.R
\name{hc_theme_ffx}
\alias{hc_theme_ffx}
\title{Firefox theme for highcharts}
\usage{
hc_theme_ffx(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Theme inspired by \url{https://www.mozilla.org/en-US/styleguide/}.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ffx())
}
