% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-flat.R
\name{hc_theme_flat}
\alias{hc_theme_flat}
\title{Flat theme for highcharts}
\usage{
hc_theme_flat(...)
}
\description{
Flat and flatdark theme is inspired by
\url{https://github.com/chriskempson/base16} and
\url{https://github.com/cttobin/ggthemr#flat}
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_flat())
}
