% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-ggplot2.R
\name{hc_theme_ggplot2}
\alias{hc_theme_ggplot2}
\title{ggplot2 theme for highcharts}
\usage{
hc_theme_ggplot2(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Based on \url{https://ggplot2.tidyverse.org/}.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_ggplot2())
}
