% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-gridlight.R
\name{hc_theme_gridlight}
\alias{hc_theme_gridlight}
\title{Grid Light theme for highchart}
\usage{
hc_theme_gridlight()
}
\description{
Grid Light theme for highchart
}
\examples{

highchart() \%>\% 
  hc_add_series(data = c(7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2,
                        26.5, 23.3, 18.3, 13.9, 9.6),
                        type = "column") \%>\% 
  hc_add_theme(hc_theme_gridlight())

}

