% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-handdrawn.R
\name{hc_theme_handdrawn}
\alias{hc_theme_handdrawn}
\title{Hand Drawn theme for highcharts}
\usage{
hc_theme_handdrawn()
}
\description{
Hand Drawn theme for highcharts. Inspirated by https://www.amcharts.com/inspiration/hand-drawn/
}
\examples{

highchart() \%>\% 
  hc_add_series(data = c(7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2,
                        26.5, 23.3, 18.3, 13.9, 9.6),
                        type = "column", color = "black") \%>\% 
  hc_add_theme(hc_theme_handdrawn())

}

