% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-handdrawn.R
\name{hc_theme_handdrawn}
\alias{hc_theme_handdrawn}
\title{Hand Drawn theme for highcharts}
\usage{
hc_theme_handdrawn(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Hand Drawn theme for highcharts. Inspired by
\url{https://www.amcharts.com/inspiration/hand-drawn/}.
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_handdrawn())

}
