% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{hc_theme_merge}
\alias{hc_theme_merge}
\title{Merge themes}
\usage{
hc_theme_merge(...)
}
\arguments{
\item{...}{\code{hc_theme} objects.}
}
\description{
Function to combine hc_theme objects.
}
\examples{

thm <- hc_theme_merge(
  hc_theme_darkunica(),
  hc_theme(
    chart = list(
      backgroundColor = "transparent",
      divBackgroundImage = "http://cdn.wall-pix.net/albums/art-3Dview/00025095.jpg"
    ),
    title = list(
      style = list(
        color = 'white',
        fontFamily = "Erica One"
      )
    )
  )
)

}

