% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-monokai.R
\name{hc_theme_monokai}
\alias{hc_theme_monokai}
\title{Monokai theme for highcharts}
\usage{
hc_theme_monokai(...)
}
\arguments{
\item{...}{A named parameters to modify the theme.}
}
\description{
Monokai is a well know text editor theme.
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_monokai())
  
}
