% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-null.R
\name{hc_theme_null}
\alias{hc_theme_null}
\title{Null theme for highcharts}
\usage{
hc_theme_null(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Null theme for highcharts. Axis are removed (\code{visible = FALSE}).
}
\examples{

highcharts_demo() \%>\%
  hc_add_theme(hc_theme_null())
}
