% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smpl.R
\name{hc_theme_smpl}
\alias{hc_theme_smpl}
\title{Simple Theme}
\usage{
hc_theme_smpl(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Desing inspired by \url{https://github.com/hrbrmstr/hrbrmisc/blob/master/R/ggplot.r}
and color by \url{https://www.materialui.co/colors}
}
\examples{

hc_demo() \%>\% 
  hc_add_theme(hc_theme_smpl())

}

