% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-superheroes.R
\name{hc_theme_superheroes}
\alias{hc_theme_superheroes}
\title{Superheroes theme for highcharts}
\usage{
hc_theme_superheroes(...)
}
\arguments{
\item{...}{Named argument to modify the theme}
}
\description{
Inspired by \url{https://public.tableau.com/profile/ryansmith#!/vizhome/HeroesofNewYork/SuperheroesinNewYork}
}
\examples{

highcharts_demo() \%>\% 
  hc_add_theme(hc_theme_superheroes())

}
