% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{hc_title}
\alias{hc_title}
\title{Adding titles options to highchart object}
\usage{
hc_title(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#title}
}

