% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart-api.R
\name{hc_tooltip}
\alias{hc_tooltip}
\title{Adding tooltip options to highchart object}
\usage{
hc_tooltip(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in \url{http://api.highcharts.com/highcharts}.}
}
\description{
Arguments are defined in \url{http://api.highcharts.com/highcharts#tooltip}
}

