% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts-api-helpers.R, R/highcharts-api.R
\name{hc_yAxis_multiples}
\alias{hc_yAxis_multiples}
\alias{hc_xAxis}
\title{xAxis options for highcharter objects}
\usage{
hc_yAxis_multiples(hc, ...)

hc_xAxis(hc, ...)
}
\arguments{
\item{hc}{A \code{highchart} \code{htmlwidget} object.}

\item{...}{Arguments defined in https://api.highcharts.com/highcharts/xAxis.}
}
\description{
The X axis or category axis. Normally this is the horizontal axis,
though if the chart is inverted this is the vertical axis. In case of
multiple axes, the xAxis node is an array of configuration objects.
See the Axis class for programmatic
access to the axis.
}
\details{
In Highmaps, the axis is hidden, but it is used behind the scenes to
control features like zooming and panning. Zooming is in effect the same
as setting the extremes of one of the exes.
}
\examples{

highchart() \%>\%
  hc_add_series(
    data = c(7.0, 6.9, 9.5, 14.5, 18.2, 21.5, 25.2, 26.5, 23.3, 18.3, 13.9, 9.6),
    type = "spline"
    ) \%>\% 
  hc_xAxis(
    title = list(text = "x Axis at top"),
    alternateGridColor = "#FDFFD5",
    opposite = TRUE,
    plotLines = list(
      list(
        label = list(text = "This is a plotLine"),
        color = "#FF0000",
        width = 2,
        value = 5.5
        )
      )
    )

}
