% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcbar}
\alias{hcbar}
\title{Function to make a bar chart}
\usage{
hcbar(x, ...)
}
\arguments{
\item{x}{A character o factor vector.}

\item{...}{Aditional shared arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make a bar chart
}

