% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcbar}
\alias{hcbar}
\title{Shortcut to make a bar chart}
\usage{
hcbar(x, ...)
}
\arguments{
\item{x}{A character or factor vector.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make a bar chart
}

