% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcboxplot}
\alias{hcboxplot}
\title{Shortcut to make a boxplot}
\usage{
hcboxplot(x = NULL, var = NULL, var2 = NULL, outliers = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{var}{A string vector same length of x.}

\item{var2}{A string vector same length of x.}

\item{outliers}{A boolean value to show or not the outliers.}

\item{...}{Additional arguments for the data series \url{https://api.highcharts.com/highcharts/series}.}
}
\description{
Shortcut to make a boxplot
}
\examples{
\dontrun{
hcboxplot(x = iris$Sepal.Length, var = iris$Species, color = "red")
}
}
