% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcboxplot}
\alias{hcboxplot}
\title{Function to make a boxplot}
\usage{
hcboxplot(x, by = NULL, outliers = TRUE, horizontal = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{by}{A string vector same length of x.}

\item{outliers}{A boolean value to show or not the outliers.}

\item{horizontal}{A boolean value indicating if the boxplots should be horizontal.}

\item{...}{Aditional arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make a boxplot
}

