% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcdensity}
\alias{hcdensity}
\title{Shortcut to make density charts}
\usage{
hcdensity(x, ...)
}
\arguments{
\item{x}{A numeric vector or a density object.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make density charts
}
