% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart.R
\name{hchart}
\alias{hchart}
\title{Create a highchart object from a particular data type}
\usage{
hchart(object, ...)
}
\arguments{
\item{object}{A R object.}

\item{...}{Aditional arguments for the data series 
(\url{http://api.highcharts.com/highcharts#series}).}
}
\description{
\code{hchart} uses \code{highchart} to draw a particular plot for an 
object of a particular class in a single command. This defines the S3 
generic that other classes and packages can extend.
}
\details{
Run \code{methods(hchart)} to see what objects are supported.
}

