% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart_waffle.R
\name{hchart_waffle}
\alias{hchart_waffle}
\title{Function to create  waffle charts}
\usage{
hchart_waffle(labels, counts, rows = NULL, icon = NULL, size = NULL)
}
\arguments{
\item{labels}{A character vector}

\item{counts}{A integer vector}

\item{rows}{A integer to set}

\item{icon}{A string}

\item{size}{Font size}
}
\description{
Function to create  waffle charts
}

