% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hchist}
\alias{hchist}
\title{Function to make an histogram}
\usage{
hchist(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Aditional arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make an histogram
}

