% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hchist}
\alias{hchist}
\title{Shortcut to make an histogram}
\usage{
hchist(x, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make an histogram
}
