% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcpie}
\alias{hcpie}
\title{Shorcut to make a pie chart}
\usage{
hcpie(x, ...)
}
\arguments{
\item{x}{A character o factor vector.}

\item{...}{Aditional arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Shorcut to make a pie chart
}

