% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcpie}
\alias{hcpie}
\title{Function to make a pie chart}
\usage{
hcpie(x, ...)
}
\arguments{
\item{x}{A character o factor vector.}

\item{...}{Aditional shared arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make a pie chart
}

