% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcspark}
\alias{hcspark}
\title{Shortcut to make spkarlines}
\usage{
hcspark(x = NULL, type = NULL, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{type}{Type sparkline: line, bar, etc.}

\item{...}{Additional arguments for the data series \url{http://api.highcharts.com/highcharts#series}.}
}
\description{
Shortcut to make spkarlines
}
\examples{

set.seed(123)
x <- cumsum(rnorm(10))

hcspark(x) 
hcspark(x, "columnn")
hcspark(c(1, 4, 5), "pie")
hcspark(x, type = "area")
   
}

