% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc-charts.R
\name{hcts}
\alias{hcts}
\title{Function to make time or line charts}
\usage{
hcts(x, ...)
}
\arguments{
\item{x}{A numeric vector or a time series object.}

\item{...}{Aditional shared arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Function to make time or line charts
}

