% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchart-shorcuts.R
\name{hcts}
\alias{hcts}
\title{Shorcut to make time series or line charts}
\usage{
hcts(x, ...)
}
\arguments{
\item{x}{A numeric vector or a time series object.}

\item{...}{Aditional arguments for the data series (http://api.highcharts.com/highcharts#series).}
}
\description{
Shorcut to make time series or line charts
}

