% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-package.R
\docType{package}
\name{hicharter}
\alias{hicharter}
\alias{hicharter-package}
\title{An \code{htmlwidget} interface to the
Highcharts javascript chart library}
\description{
Highcarts \url{http://www.highcharts.com/} is a mature javascript
charting library. Highcharts provide a various type of charts, from
scatters to heatmaps or treemaps.
}
\author{
Joshua Kunst (@jbkunst)
}

