% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-package.R
\name{highcarter-exports}
\alias{JS}
\alias{\%>\%}
\alias{highcarter-exports}
\title{highcharter exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the highcarter package to avoid listing the magrittr
as Depends of highcarters.
}

