% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{highchart}
\alias{highchart}
\title{Create a Highcharts chart widget}
\usage{
highchart(hc_opts = list(), theme = NULL, width = NULL, height = NULL,
  debug = FALSE)
}
\arguments{
\item{hc_opts}{A \code{list} object containing options defined as 
\url{http://api.highcharts.com/highcharts}.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{debug}{A boolean value if you want to print in the browser console the 
parameters given to \code{highchart}.}
}
\description{
This function creates a Highchart chart using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}

