% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{highchart2}
\alias{highchart2}
\title{Create a Highcharts chart widget}
\usage{
highchart2(hc_opts = list(), theme = NULL, width = NULL, height = NULL,
  elementId = NULL, debug = FALSE)
}
\arguments{
\item{hc_opts}{A \code{list} object containing options defined as 
\url{http://api.highcharts.com/highcharts}.}

\item{theme}{A \code{hc_theme} class object}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}

\item{elementId}{Use an explicit element ID for the widget.}

\item{debug}{A boolean value if you want to print in the browser console the 
parameters given to \code{highchart}.}
}
\description{
This widgets don't support options yet.
}
\details{
This function creates a Highchart chart using \pkg{htmlwidgets}. The
widget can be rendered on HTML pages generated from R Markdown, Shiny, or
other applications.
}

