% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter-package.R
\name{highcharter-exports}
\alias{highcharter-exports}
\alias{JS}
\alias{tags}
\title{highcharter exported operators and S3 methods}
\description{
The following functions are imported and then re-exported
from the highcharter package to avoid listing the magrittr
as Depends of highcharter.
}
