\name{highcharter-package}
\alias{highcharter-package}
\alias{highcharter}
\docType{package}
\title{
\packageTitle{highcharter}
}
\description{
\packageDescription{highcharter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{highcharter}
\packageIndices{highcharter}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{highcharter}

Maintainer: \packageMaintainer{highcharter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
