% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlwidget.R
\name{highchartOutput}
\alias{highchartOutput}
\title{Widget output function for use in Shiny}
\usage{
highchartOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{The name of the input.}

\item{width}{A numeric input in pixels.}

\item{height}{A numeric input in pixels.}
}
\description{
Widget output function for use in Shiny
}

