% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw_grid.R
\name{hw_grid}
\alias{hw_grid}
\title{Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}}
\usage{
hw_grid(..., ncol = NULL, rowheight = NULL)
}
\arguments{
\item{...}{either individual \code{highchart} objects or a mixture of
    individual \code{highchart} objects
and \code{list}s of \code{highchart} objects.}

\item{ncol}{how many columns in the grid}
}
\description{
Lays out highchart widgets into a "grid", similar to
\code{grid.arrange} from \code{gridExtra}
}

