% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pokemon.R
\docType{data}
\name{pokemon}
\alias{pokemon}
\title{pokemon}
\format{A \code{data frame} with 718 observations and 20 variables.}
\source{
\url{jkunst.com/r/pokemon-visualize-em-all/}
}
\usage{
pokemon
}
\description{
Information about 718 pokemon.
}
\section{Variables}{


\itemize{
 \item \code{id}:
 \item \code{pokemon}:
 \item \code{species_id}:
 \item \code{height}:
 \item \code{weight}:
 \item \code{base_experience}:
 \item \code{type_1}:
 \item \code{type_2}:
 \item \code{attack}:
 \item \code{defense}:
 \item \code{hp}:
 \item \code{special_attack}:
 \item \code{special_defense}:
 \item \code{speed}:
 \item \code{color_1}:
 \item \code{color_2}:
 \item \code{color_f}:
 \item \code{egg_group_1}:
 \item \code{egg_group_2}:
 \item \code{url_image}:
}
}
\keyword{datasets}

