% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highchart.R
\name{renderHighchart}
\alias{renderHighchart}
\title{Widget render function for use in Shiny}
\usage{
renderHighchart(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

