% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharter.R
\name{renderHighchart}
\alias{renderHighchart}
\alias{renderHighchartZ}
\title{Widget render function for use in Shiny}
\usage{
renderHighchart(expr, env = parent.frame(), quoted = FALSE)

renderHighchartZ(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{A highchart expression.}

\item{env}{A environment.}

\item{quoted}{A boolean value.}
}
\description{
Widget render function for use in Shiny
}
