% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tooltip_table}
\alias{tooltip_table}
\title{Helper for make table in tooltips}
\usage{
tooltip_table(x, y, title = NULL, img = NULL, ...)
}
\arguments{
\item{x}{A string vector with description text}

\item{y}{A string with accesors ex: \code{point.series.name},
\code{point.x}}

\item{title}{A title tag with accessor or string}

\item{img}{Img tag}

\item{...}{html attributes for the table element}
}
\description{
Helper to make table in tooltips for the \code{pointFormat}
parameter in \code{hc_tooltip}
}
\examples{

x <- c("Income:", "Genre", "Runtime")
y <- c("$ {point.y}", "{point.series.options.extra.genre}",
       "{point.series.options.extra.runtime}")

tooltip_table(x, y)

}

